==> Analysis/02-calculate-user-error.Rmd <==
---
title: "02-calculate-user-error"
author: "Ariel Marcy"
date: "2018-09-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Calculate user error
Geometric morphometric studies require a lot of user-generated data and it's important to assess how much user error is created during landmarking. 

First, we visualized whether any outliers can be explained by mistakes in landmark placement (e.g. a point "falling through" the nasal to land on the inner basicranium).

Second, we compare the shape variation between 20 individuals, from a variety of different species, landmarked twice to calculate the intra-observer (i.e. digitizer) error.

Third, we'll finalize the specimens to be included in future analyses. 

### Load packages, functions, and data from previous steps
```{r message = FALSE}
library(stringr)
library(stringi)
library(data.table)
library(geomorph)
source("../Functions/utilities.R")  # custom functions defined in utilities.R
load(file = "../Data/Processed/01-bilat-data-w-ERR.rda")  # data created by the first script
```

## 1) Check for outliers and fix landmarking errors
_geomorph_ comes with a handy function to check for outliers. Since our sample has a wide range of species, we decided to include the groups factor. This means the `plotOutliers` returns likely outliers for each genus instead of the group as a whole. When groups were not defined, only large species were returned as "outliers". 
```{r}
# Plot outliers by genus for small patch dataset
outliers <- plotOutliers(shape, groups = sp.info$Genus, inspect.outliers = T)
```

### Visualize why they are outliers by looking at landmark vectors
We can examine the landmark placements in outliers as compared to the specimen closest to the median, based on the plots above. This allows us to determine if there were any major landmarking errors that might impact further analyses.

Genus abbreviatons and number of outliers were inserted where needed for each genus. This code was only used once: after all landmarks were captured and before the other major analyses were completed.

We wrote a function, `MatchSpecShape()` to help us visualize outlier shape information. The function returns the 3D shape of specimen of interest. We can use it to compare the outliers to the median specimen in its genus.
```{r}
# Get shape data for outliers indicated by plots above
out.1 <- MatchSpecShape(outliers$Mel[1], sp.info, shape)  # 1 = 1st outlier in genus Mel

# Find median specimen of group with outlier
med.spec <- outliers$Mel[median(outliers$Mel)]  # median specimen in Mel
med.shape <- MatchSpecShape(med.spec, sp.info, shape)

# Compare landmark locations in outlier specimens to median specimen
plotRefToTarget(med.shape, out.1, method = "vector", label = FALSE)
```
**Repeat the above for each genus/patch protocol as needed. When landmarking errors are found, fix in Viewbox, export new coordinates, and begin again from script 01.** 

### ID outliers within genus *Pseudomys*
To ID outliers within each species of the speciose genus *Pseudomys*, we needed to create shape and metadata files just for them so the `plotOutliers()` function can separate them. Otherwise only individuals from P oralis show up as outliers. 
```{r}
# Isolate Pseudomys species
is.Pse <- which(sp.info$Genus == "Pse")
Pse.sm.shape <- shape[, , is.Pse]  # shape
Pse.sm.info <- droplevels(sp.info[is.Pse, ])  # metadata

# Run plotOutliers
outliers.Pse <- plotOutliers(Pse.sm.shape, Pse.sm.info$Species, inspect.outliers = TRUE)
```

## 2) Calculate intra-observer error
Now that the shape data have been processed via GPA + bilateral symmetry and corrected for obvious landmarking errors, we can separate the data intended to measure user error from the main data for the remaining analyses. 

First, create an error measurement dataset for both protocols that includes replicates and the specimens they duplicated (i.e. each CatNum should have two entries, one labeled "ERR", and one labled "non"). These data will be used to calculate user error. 
```{r}
# Find index for both versions of specimens with replicates using rep.spec list made in script 01.
rep.CatNums <- as.character(str_sub(rep.spec, 9, -1))  # isolate CatNum
error.data <- which(sp.info$CatNum %in% rep.CatNums)

# Subset shape and metadata tables
err.shape <- shape[, , error.data]
err.info <- sp.info[error.data, ]
```

### Calculate user error aka Repeatability
Here, we calculate repeatability for each patch protocol from formulas 1-3 in Fruciano 2016. The function `RepAbility()` was modified from Marcy et al. 2018.
```{r}
RepAbility(err.shape, err.info$CatNum, n.Rep = 2)  # note: n.Rep = 2 in this case because we measured each specimen twice
```

### Calculate repeatability without patch points
Patch points could artificially inflate our repeatability numbers since they are placed automatically by the Viewbox software. To test this, we calculated repeatability with patch points removed. 
```{r}
# Find patch numbers
patches <- str_detect(pt.names, "PAT")
pat.num <- which(patches == TRUE)

# Subset the shape datasets for only fixed LMs and curve semi-LMs
patchless.shape <- err.shape[-pat.num, , ]

# Re-run repeatability
RepAbility(patchless.shape, err.info$CatNum, n.Rep = 2)
```
Taking out patch points decreases the repeatability by ~1.4%.

## 3) Make main dataset
Make the datasets without replicates to be used for the rest of the analyses.
```{r}
main <- which(sp.info$Rep == "non")
shape <- shape[, , main]
sp.info <- sp.info[main, ]
```

### Remove outliers from dataset
Three specimens had major landmarking errors: landmarks consistently placed in the wrong places. Therefore, we opted to remove these specimens.
```{r}
# Find indices
rat.out <- which(sp.info$CatNum == "C11381")  # Rattus norvegicus
uro.out <- which(sp.info$CatNum == "J17610")  # Uromys caudimaculatus

# Remove from datasets
info <- sp.info[-c(rat.out, uro.out), ]  # metadata
shape <- shape[, , -c(rat.out, uro.out)]  # shape data
```

### Save intermediate data
```{r}
save(shape, info, file = "../Data/Processed/02-main-data.rda")
```
