==> Analysis/03-measure-static-allometry.Rmd <==
---
title: "03-measure-static-allometry"
author: "Ariel Marcy"
date: "2019-01-11"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Static Allometry
Changing size is one of the most common ways that organisms can change their shape. Modifications to growth during development often have a profound impact on adult shape. The tests in this script detect how much size appears to drive shape change in our sample.

This script generates Figure 1 and data for Table 1a. 

It also produces supplementary Figures TK 

### Load packages, functions, and data
```{r message = FALSE}
library(stringr)
library(data.table)
library(geomorph)
library(RRPP)  # needed for pairwise comparisons with new `geomorph`
source("../Functions/utilities.R")  # loads custom functions
load(file = "../Data/Processed/02-main-data.rda")
load(file = "../Data/Processed/03-color-data.rda") # TEMP
load(file = "../Data/Processed/04-phylo-data.rda") # TEMP
```

## 1) Measure static allometry: Homogeneity of Slopes Test by species
Isolate only those species with at least 8 specimens.
```{r}
spec.counts <- as.data.frame(table(info$Taxa))
to.remove <- spec.counts[which(spec.counts$Freq <= 7), ]
species.rm <- as.vector(to.remove$Var1)

# For loop to take out these specimens
dataset.rm <- NULL  # initiate blank data frame
for(n in 1:dim(to.remove)[1]) {
        dataset.rm <- c(dataset.rm, which(info$Taxa == species.rm[n]))
}

# Remove species from:
info.HOS <- info[c(-dataset.rm), ]  # Metadata
shape.HOS <- shape[, , c(-sort(dataset.rm))]  # Shape dataset
```

Run allometry tests as well as a homogeneity of slopes test. This test may take 1-2 min to run. Updated for `geomorph` 3.1.0. 
```{r}
# HOS test Defined as a function for conciseness in the next block. Returns pairwise object form RRPP
TestHOS <- function(shape, size, spp, iter = 500, lm.fun) {
        # Data frame
        spp.gdf <- geomorph.data.frame(coords = shape, size = size, spp = spp)
        # Models
        fit.unique <- lm.fun(coords ~ log(size) * spp, iter = iter, data = spp.gdf, print.progress = FALSE)  # unique allometries by species
        fit.common <- lm.fun(coords ~ log(size) + spp, iter = iter, data = spp.gdf, print.progress = FALSE)  # common allometries across species
        
        anova(fit.common, fit.unique, print.progress = FALSE)

        return(pairwise(fit.common, fit.unique, groups = spp.gdf$spp, print.progress = FALSE))  # includes Homogeneity of Slope pairwise comparisons
}

# Run with HOS dataset sampled above
pw.comp <- TestHOS(shape.HOS, info.HOS$Csize, info.HOS$Taxa, lm.fun = procD.lm)  

# Count number of significant pairwise comparisons of alloemtric slope
pw.summary <- summary.pairwise(pw.comp, test.type = "VC", angle.type = "deg", confidence = 0.95, show.vectors = TRUE)  
actual.pw <- length(which(pw.summary.2$pairwise.tables$P < 0.05))

# Export p-value matrix to Results folder
write.csv(pw.summary$pairwise.tables$P, "../Data/Results/TableS4_HOS.csv")
```

## Randomize specimens to get NULL expantancy for pairwise significant results
We'll do the randomization procedure 100 times. This takes at least 20 minutes to run.
```{r}
# Define function to shuffle species attributions and run PW comparisons; this randomizes species relative to centroid size
null.pairwise <- function(shape, info, iter = 500, confidence = 0.95) {
        rand.spp <- sample(info$Taxa)
        
        # Run the test with random sample
        pw.comp <- TestHOS(shape, info$Csize, rand.spp, lm.fun = procD.lm)
        pw.summary <- summary.pairwise(pw.comp, test.type = "VC", angle.type = "deg", confidence = confidence, show.vectors = TRUE)
        
        # Count number of significant pairwise comparisons of allometric slope
        alpha <- 1 - confidence
        return(length(which(pw.summary$pairwise.tables$P < alpha)))
}

# Do this 100 times
count.pw <- 1
for(i in 1:100){
        null.pw <- null.pairwise(shape = shape.HOS, info.HOS, iter = 500)
        count.pw <- c(count.pw, null.pw)
}

# Visualize with histogram
hist(count.pw[2:length(count.pw)])  # and remove initializing 1
abline(v = actual.pw, lwd = 3)  # compare random to real result above
```

## 2) Visualize Static Allometry: Predicted Value
In _geomorph_, static allometry can be visualized with log(centroid size) plotted versus the projected regression score of shape ~ size, a common way of quantifying "size-less" shape.
```{r}
# Run static allometry test with all specimens
spp.gdf <- geomorph.data.frame(shape = shape, Species = info$Taxa, size = info$Csize)
spp.results <- procD.allometry(shape ~ log(size), f2 = ~Species, iter = 500, data = spp.gdf)

# Find the mean point for each static allometry
mean.size <- aggregate(spp.results$size, list(Taxa = info$Taxa), mean)
mean.shape <- aggregate(spp.results$Reg.proj, list(Taxa = info$Taxa), mean)
```

### Plot/export static and evolutionary allometry multipanel Figure 1
Here we consolidate the information above into two static allometry plots, one with all the data points and one with just the regression lines for each species to compare with the evolutionary allometry of the entire group. 
```{r}
setEPS()  # sets up plot export
postscript("../Data/Results/Figure2_Static_Evo_Allo.eps")
layout(matrix(c(1,1,2,2), 2, 2, byrow = TRUE))

# Plot of all data points with evolutionary allometry
par(mar = c(1, 4, 1.5, 0))
plot(x = log(spp.results$size),
     y = spp.results$Reg.proj,
     xlim = c(4.7, 6.4),
     col = sp.col.gen, 
     pch = pch.spp,
     bg = sp.col.gen,
     main = "Static and evolutionary allometry",
     xlab = NA,
     ylab = "Shape (Regression Score)",
     bty = "n")
text(min(log(spp.results$size)), max(spp.results$Reg.proj), "a", cex = 2)
abline(lm(mean.shape$V1 ~ log(mean.size$x)), col = "dark grey", lwd = 2)  # plots evolutionary allometry
text(5.5, -0.04, paste("Rsq of log(size) = ", round(spp.results$aov.table$Rsq[1], 2)), cex = 1.1, col = "black")

# Species Legend
legend(6.05, 0.06, legend = phylo.names, col = col.unique.spp, border = NULL, pch = pch.unique.spp, pt.bg = col.unique.spp, cex = 0.97, pt.cex = 1, ncol = 2, bg = "white")

# Plot of static allometry with evolutionary allometry
par(mar=c(5, 4, 1.5, 0))
plot(x = log(spp.results$size),
     y = spp.results$pred.val,
     xlim = c(4.7, 6.4),
     col = sp.col.gen, 
     pch = pch.spp,
     bg = sp.col.gen,
     xlab = "Log(centroid size)",
     ylab = "Shape (Predicted Value)",
     bty = "n")
text(min(log(spp.results$size)), max(spp.results$pred.val), "b", cex = 2)
text(5.6, -0.08, paste("Rsq of log(size):species = ", round(spp.results$aov.table$Rsq[3], 2)), cex = 1.1, col = "black")

dev.off()
```

### Save intermediate data
```{r}
save(spp.results, file = "../Data/Processed/04-allometry-data.rda")
```


## Perform rarefaction along the tree
This addresses reviewer concerns over small sample sizes in each species bin and whether it underlies some or all of our results showing parallel slopes across almost all species. 

The code below is based on Dr Thomas Guillerme's [landvR vignette here](https://github.com/TGuillerme/landvR/blob/master/inst/vignettes/rarefy_regressions.Rmd)

### Getting each clade
First we need to identify and isolate each element from each partition (clade) using a modified version from the `ape::prop.part` function.
```{r}
# Create simple data table with label, centroid size, and RegProj
short.label <- function(label) paste0(substr(strsplit(label, split = "_")[[1]][1], 1, 1), "_", substr(strsplit(label, split = "_")[[1]][2], 1, 3))
aus.tree.stlbl <- aus.tree
aus.tree.stlbl$tip.label <- unname(sapply(aus.tree.stlbl$tip.label, short.label))

rare.table <- data.frame(info$Taxa, log(info$Csize), unname(spp.results$Reg.proj))
colnames(rare.table) <- c("Taxa", "Csize", "RegProj")
rownames(rare.table) <- NULL

# Modify a function from ape to ID and isolate each element from each clade
prop.part.names <- function(phy, singletons = FALSE) {
    # Sanitizing
    if(class(phy) != "phylo") {
        stop("phy must be a \"phylo\" object.")
    }
    if(class(singletons) != "logical") {
        stop("singletons must be logical.")
    }

    # Get the bipartitions
    clades <- ape::prop.part(phy)

    # Get the tips names for each clades
    clades <- lapply(clades, function(clade, labels) labels[clade], labels = attr(clades, "labels"))

    # Add the tip names
    if(singletons) {
        clades <- c(clades, as.list(phy$tip.label))
    }

    return(clades)
}

# Getting the tree partitions (with singletons)
clades <- prop.part.names(aus.tree.stlbl, singletons = TRUE)

# Make function for getting the dataset for each clade
get.data.clade <- function(clade, data, species.col.n, data.cols) {
    return(data[which(data[, species.col.n] %in% clade), data.cols])
}

# Use function to get all the datasets per clades
data_clades <- lapply(clades, get.data.clade, data = rare.table,
                      species.col.n = 1, data.cols = c(2:3))
```

### Applying the tests
First we need to define the function that will get us the statistic. In our case we want the slope from a linear model (but any statistic can be used):

```{r}
# Function for getting the slope of a linear model
GetSlope <- function(data) {
    return(lm(data)$coefficients[[2]])
}
```

We can then calculate the delta slope (observed - rarefied slope) by using the lowest potential number of specimens in our dataset (5 - see above).
```{r}
# Define function for rarification or bootstrap
rarefy.stat <- function(data, stat.fun, rarefaction, replicates = 100, observed, ...) {
    # Sanitizing
    nrow_data <- nrow(data)
    if(!is.null(rarefaction)) {
        # Check if rarefaction is bigger than the # of rows in the data set
        if(rarefaction > nrow_data) {
            warning(paste0("Rarefaction value (", rarefaction, ") is bigger than the number of rows in the data (", nrow_data, ").\nThe statistic will be only bootstrapped."))
            # Set rarefaction as the number of rows in the data (bootstrap)
            rarefaction <- nrow_data    
        }
    } else {
        # Set rarefaction as the number of rows in the data (bootstrap)
        rarefaction <- nrow(data)
    }
    
    # Check the observed value
    if(!missing(observed)) {
    } else {
        observed <- stat.fun(data)
    }

    # Rarefaction function
    results <- replicate(replicates, stat.fun(data[sample(1:nrow_data, rarefaction, replace = TRUE), ]))

    # Getting the differences
    return(results - observed)
}

# Getting the rarefied statistics for each clade
delta_slopes <- lapply(data_clades, rarefy.stat, stat.fun = GetSlope, rarefaction = 5)
```

We can also compare them to the slope of the the full dataset:
```{r}
# Getting the rarefied statistics for each clade
delta_slopes_overall <- lapply(data_clades, rarefy.stat, stat.fun = GetSlope, rarefaction = 5, observed = GetSlope(rare.table[, c(2, 3)]))
```

### Summarising the results
We can then summarise these results per clade size (that's already done in this example) and adding the clade names (if they exist):

```{r}
# Adding the names to the slopes
names(delta_slopes) <- unlist(lapply(clades, paste, collapse = "+"))
names(delta_slopes_overall) <- unlist(lapply(clades, paste, collapse = "+"))
```

And then transform these statistics into tables to simply plot them with `boxplot`:
```{r, fig.width = 12, fig.height = 12}
par(mfrow = c(2, 1))
# Combining the results into tables and plotting them
boxplot(do.call(cbind, delta_slopes),
        ylab = expression(paste(Delta, "slope")), xlab = "Clade",
        main = "Slope difference for each clade", las = 2)
abline(h = 0, lty = 3)
boxplot(do.call(cbind, delta_slopes_overall),
        ylab = expression(paste(Delta, "slope")), xlab = "Clade",
        main = "Slope difference for each clade (compared to the whole dataset)", las = 2)
abline(h = 0, lty = 3)  # since observered minus expected should be 0
```

## Checking the validity of the results above by randomizing species into "fake clades"
First create some fake clade groupings (i.e. non phylogenetic) and connect it to real data by individuals
```{r}
# Getting the same number of "groups" per "clades"
clades <- prop.part.names(aus.tree)
n.spp.group <- sort(unlist(lapply(clades, length)), decreasing = TRUE)  # sort by big->small for graphing at the end

# Make function to get rarefied statistics for a fake clade grouping
RareFlakes <- function(n.spp.group, rare.table) {
        # Spp names for each fake clade as per n in each clade
        fake.clades <- lapply(n.spp.group, function(n.spp.group, names = rare.table$Taxa) sample(names, n.spp.group))
        
        # Use get.data.clade function from above to grab datasets for clades
        data.flakes <- lapply(fake.clades, get.data.clade, data = rare.table, species.col.n = 1, data.cols = c(2:3))
        
        # Get the rarefied statistics for each clade
        delta.slopes <- lapply(data.flakes, rarefy.stat, stat.fun = GetSlope, rarefaction = 5)
        
        return(delta.slopes)
}
```

### Applying the tests to the "fake clades" (called data.flakes) x 100
Again, calculate the delta slope (observed - rarefied slope) by using the lowest potential number of specimens in our dataset (5 - see above).
```{r}
delta.slopes <- list()
for(i in 1:100) {  # Rep 100 times
        delta.slopes[[i]] <- do.call(cbind, RareFlakes(n.spp.group, rare.table))
}
fake.results.table <- do.call(rbind, delta.slopes)  # full matrix of slope values
colnames(fake.results.table) <- n.spp.group
```

We can also compare them to the slope of the the full dataset:
```{r}
# Getting the rarefied statistics for each clade
delta.slopes.overall <- lapply(data.flakes, rarefy.stat, stat.fun = GetSlope, rarefaction = 5, observed = GetSlope(rare.table[, c(2, 3)]))
```

### Summarising the results
We can then summarise these results per clade size (that's already done in this example) and adding the clade names (if they exist):
```{r}
# Adding the names to the slopes
names(delta.slopes) <- n.spp.group
names(delta.slopes.overall) <- n.spp.group
```

And then transform these statistics into tables to simply plot them with `boxplot`:
```{r, fig.width = 12, fig.height = 12}
# An empty plot first
fake.results.list <- apply(fake.results.table, 2, list)
colnames(fake.results.table) <- n.spp.group
plot(NULL, xlim = c(0, length(fake.results.list)), ylim = quantile(unlist(fake.results.table), probs = c(0.01, 0.99)))

boxplot(fake.results.table, border = "white", ylim = quantile(unlist(fake.results.table), probs = c(0.01, 0.99)), ylab = "whatever", xlab = "Number of species")

     
# Adding the results
plot.CI(unlist(fake.results.list, recursive = FALSE), type = "polygon", CI = c(50, 95), cent.tend = median, point.col = "black")


par(mfrow = c(2, 1))

# Combining the results into tables and plotting them
boxplot(fake.results.table,
        ylab = expression(paste(Delta, "slope")), xlab = "Number of species in fake clade", main = "Slope difference for 100 FAKE CLADEs", ylim = c(-100, 100), las = 2)
abline(h = 0, lty = 3)

boxplot(do.call(cbind, delta.slopes.overall),
        ylab = expression(paste(Delta, "slope")), xlab = "Number of species in fake clade",
        main = "Slope difference for 100 FAKE CLADEs (compared to the whole dataset)", ylim = c(-350, 125), las = 2)
abline(h = 0, lty = 3)  # since observered minus expected should be 0
```
