==> Analysis/03-prepare-data.Rmd <==
---
title: "03-prepare-data"
author: "Ariel Marcy"
date: "11/7/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

## Prepare metadata, graphing vectors, and the phylogenetic tree
These data will be needed for all subsequent analyses, tables, and figures. The rda file created at the end will be needed for all future scripts. 

### Load packages, functions, and data
```{r message = FALSE}
library(stringr)
library(data.table)
library(colorspace)  # colors
library(phylotate)  # phylogeny
library(geiger)  # phylogeny
source("../Functions/utilities.R")  # loads custom functions
load(file = "../Data/Processed/02-no-error-data.rda")  # shape data w/o error
```

## Attach all needed information to the metadata
```{r}
# Remove unnecessary columns
drops <- c("Patch", "Rep", "All")
info <- info[, !(names(info) %in% drops)]

# Add trait info from Native Mice and Rats (Breed & Ford 2007)
traits <- read.csv("../Data/Processed/in_ex_traits.csv", header = TRUE)
traits$Taxa <- paste(str_sub(traits$Genus, 1, 1), str_sub(traits$Species, 1, 3), sep = "_")  # make matching Taxa column
info <- merge(info, traits, by = "Taxa", sort = TRUE)  # merge
info <- info[order(info$Order), ]  # original order / same as shape

# Deal with Pogonomys being undescribed
info[, which(info$Species == "spp")] <- "mpp"
info[, which(info$Species == "spp")] <- "mpp"



# Rename Genus & Species columns so they have their original short column name, long form Genus and Species columns keep Genus.y and Species.y
names(info)[names(info) == "Genus.x"] <- "Genus"
names(info)[names(info) == "Species.x"] <- "Species"

# Add long form names ("Genus_species")
info$FullName <- paste(info$Genus.y, info$Species.y, sep = "_")
```

## Prepare shape dataset dimnames for phylogenetic analyses
Rename shape dataset dimnames to only "Genus_species". This is so we can re-order the shape data to match the tree later on. (We can still reference the CatNum in the metadata.)
```{r}
dimnames(shape)[[3]] <- as.list(paste(info$FullName))
```

### Handle *Mastacomys*
Since *Mastacomys* is technically within the genus *Pseudomys* (Smissen & Rowe 2018), we'll make an "Effective Genus" column in the metadata where this genus is classified under *Pseudomys* for analyses that use genera.
```{r}
info$EGenus <- info$Genus  # make Effective Genus column
info$EGenus[which(info$Genus == "Mas")] <- "Pse"  # Mas becomes Pseudomys
```

## Create colors based on phylogenetic clades in Aplin (2014)
Colors for subsequent figures are set up according to column "EGenus". Colors correspond to phylogenetics - i.e. when different ancestors arrived in Australia. Each genus gets its own color. 

We make three different vectors of color:
```{r}
# Make gradients for the genera in each clade
Uro <- sequential_hcl(5, "Greens 3")  # Uromys clade
Mes <- sequential_hcl(5, "Purples 3")  # Mesembriomys clade
Pse <- sequential_hcl(5, "Blues 3")  # Pseudomys clade
Hyd <- sequential_hcl(5, "Reds 3")  # Hydromys clade
Pog <- sequential_hcl(5, "Heat 2")  # Pogonomys clade
Inv <- sequential_hcl(5, "Light Grays")  # Mus and Rattus clades

# 1) genus color by phylogeny
genera.phylo <- c("Mus", "Rat", "Pog", "Xer", "Hyd", "Not", "Pse", "Zyz", "Leg", "Uro", "Mel", "Lep", "Mes", "Con")
col.phylo <- c(Inv[3], Inv[1], Pog[3], Hyd[1], Hyd[2], Pse[1], Pse[2], Pse[4], Pse[3], Uro[1], Uro[2], Mes[3], Mes[2], Mes[1])
names(col.phylo) <- genera.phylo 

# 2) color by alphabetical genus name
col.gen <- col.phylo[match(sort(unique(info$EGenus)), names(col.phylo))] 

# 3) color vector for all data (specimen level)
sp.col.gen <- col.phylo[match(as.factor(info$EGenus), names(col.phylo))]
sp.col.gen[which(info$Species %in% c("nor", "rat"))] <- "black"  # account for invasive Rattus
```

### Create point and color vectors for species legends in phylo order
```{r}
# Put color and point legends in phylogenetic order
phylo.names <- c("Mus_mus", "Rat_rat", "Rat_nor", "Rat_lut", "Rat_leu", "Rat_tun", "Rat_vil", "Rat_sor", "Rat_fus", "Pog_mol", "Xer_myo", "Hyd_chr", "Not_cer", "Not_fus", "Not_mit", "Not_ale", "Pse_pat", "Pse_gra", "Pse_ora", "Pse_sho", "Pse_des", "Pse_apo", "Pse_aus", "Pse_hig", "Pse_her", "Pse_nov", "Pse_del", "Pse_fus", "Zyz_arg", "Zyz_ped", "Leg_for", "Uro_cau", "Mel_cer", "Mel_cap", "Mel_bur", "Lep_con", "Mes_gou", "Con_pen")

# Points for legend
pch.unique.spp <- PointOutDiffSpp(phylo.names)  # pch by unique spp in a genus
pch.unique.spp[which(phylo.names == "Pse_fus")] <- 2  # handle Mastacomys

# Points coded for each individual specimen
info$Name <- paste(info$EGenus, info$Species, sep = "_")
phylo.order <- match(sort(unique(info$Name)), phylo.names)
pch.spp <- PlotByGroup(info, "Name", pch.unique.spp[phylo.order])  

# Make color vector for legends by species
col.taxa <- match(str_sub(phylo.names, 1, 3), names(col.gen))
col.unique.spp <- col.gen[col.taxa]
col.unique.spp[which(phylo.names %in% c("Rat_nor", "Rat_rat"))] <- "black"  # account for invasive Rattus

# Give short names back to phylo.names
phylo.names <- paste(str_sub(phylo.names, 1, 1), str_sub(phylo.names, 5, 7))
phylo.names[which(phylo.names == "P fus")] <- "M fus"  # handle Mastacomys
```

# Prepare data for phylogenetic analyses
First, the shape data have to be averaged by species. This step has to be done before plotting evolutionary allometry and working with the phylogeny. Second, the phylogeny has to be trimmed to the species in the shape data as the commands in `geomorph` require the tree and datasets to have exactly the same set of species.

## 1) Calculate mean shape and centroid size for each species
The mean shape function, `mshape()` outputs a 2D array. Therefore, the `for` loop below stores mean shapes in a 2D array format recognized by `geomorph`: (n x [p x k]), where n is number of specimens, p is number of landmark points, and k is number of dimensions. Finally, we convert to a 3D array with format (p x k x n) since a 3D array is required by most `geomorph` functions. 
```{r}
# For loop to subset shape data by species and calculate mean shape
mean.shapes.2D <- NULL  # initiate blank data frame
full.names <- unique(info$FullName)  # list of species

for (i in 1:length(full.names)) {  
        is.taxa <- which(info$FullName == full.names[i])
        shape.for.mean.2D <- shape[, , is.taxa]
        mean.shapes.2D <- rbind(mean.shapes.2D, mshape(shape.for.mean.2D))  # output cols must = 3 for 3D
}

mean.shapes <- arrayspecs(mean.shapes.2D, 325, 3)  # convert to 3D array with 325 landmarks and 3 for 3D landmarks in this study
```

Create metadata and put mean shape and metadata in alphabetical order.
```{r}
# Alphabetical order for mean shape
dimnames(mean.shapes)[[3]] <- full.names  # Genus_species format
mean.shapes <- mean.shapes[, , sort(dimnames(mean.shapes)[[3]])]  # ABC order

# Alphabetical order for mean shape metadata
info.means <- info[!duplicated(info$Taxa), ]  # make metadata
drops <- c("CatNum", "Csize", "Order")  # remove outdated columns
info.means <- info.means[, !(names(info.means) %in% drops)]
info.means <- info.means[order(info.means$FullName), ]  # ABC order
```

### Calculate mean centroid size
We need to update the centoid size column for info.means to contain the *mean* centroid size by taxa. Centroid size for all 317 specimens is stored in the info.phylo metadata table.
```{r}
# Calculate mean centroid size from original Csize data and store in info.means
cs.means <- aggregate(info[, which(colnames(info) == "Csize")], list(Taxa = info$FullName), mean)  # aggregate also alphabetizes it
info.means$MeanCsize <- cs.means[, 2]  # store in dataframe
```

## 2) Prepare phylogenetic tree
The most recent tree for Australian rodents is from Smissen & Rowe 2018. The tree is a fossil calibrated ultrametric tree from BEAST2 for Hydromyini, which includes all the Australo-Papuan old endemics in Murinae. This tree adds 6 more species from Australia to the Smissen & Rowe (2018) phylogeny. 
```{r}
raw.tree <- read_annotated("../Data/Processed/Marcy-BEAST01.con.tre")
```

### Extract only Genus and Species names from the tree
The tree file has uneven amounts of information for each entry, some entries have catalog numbers and others do not. Therefore we needed a `for` loop to grab the first two elements from a list, which always corresponded to Genus and Species.

We reassign only Genus and Species names (in the form "Genus_species") to the tree so that it is easier to compare to the shape data. 
```{r}
no.cat.tree <- raw.tree  # initiate new, modified tree with no CatNums
label.list <- strsplit(raw.tree$tip.label, "_")  # separate info into list

# For loop finds Genus & Species and makes tip name "Genus_species"
for (i in 1:length(label.list)) {
        label <- unlist(label.list[i])
        no.cat.tree$tip.label[i] <- paste(label[1], label[2], sep = "_")
}
```

### Re-label nodes
Since the 2018 Smissen & Rowe tree focused primarily on New Guinea species, it does not have all of the Australian species in our shape dataset. However, it does include sister species with the same relative branch length to other genera in the study. Therefore, we decided to re-label one of the nodes to the sister Australian species (currently undescribed, but this will be encoded later in script 08). 

**Relabeled node:**
_Pogonomys loriae_ (NG) to _Pogonomys spp_ (Aus)
```{r}
relabel.tree <- no.cat.tree
P.lor <- which(str_detect(no.cat.tree$tip.label, "Pogonomys_loriae"))
relabel.tree$tip.label[P.lor] <- paste("Pogonomys_mollipilosus")
```

**Correct Rattus**
R. lutreolus and R. leucopus were switched relative to Rowe et al 2011 and other recent phylogenies. Likely due to a simple labling error in phylogenetic preparation. Here we correct it. 
```{r}
R.lut <- which(str_detect(relabel.tree$tip.label, "Rattus_lutreolus"))
R.leu <- which(str_detect(relabel.tree$tip.label, "Rattus_leucopus"))

relabel.tree$tip.label[R.lut] <- paste("Rattus_leucopus")
relabel.tree$tip.label[R.leu] <- paste("Rattus_lutreolus")
```

### Prune the tree and shape datasets to match each other
Now that the names in the Smissen & Rowe tree can be compared to our info.phylo metadata, we can prune both datasets to have the same species.
```{r}
# First, prune the tree of species not in the shape data
concord <- name.check(relabel.tree, relabel.tree$tip.label, dimnames(mean.shapes)[[3]])  # check concordance
aus.tree <- drop.tip(relabel.tree, concord$tree_not_data)  # prune tree to Aus sample
```

Next, subset and re-order the metadata and shape datasets to match the order of species found in the tree. 
```{r}
# Subsetting
to.remove <- name.check(aus.tree, aus.tree$tip.label, dimnames(mean.shapes)[[3]])
info.means.tree <- info.means[!is.element(info.means$FullName, to.remove$data_not_tree), ]  # metadata
mean.shapes.tree <- mean.shapes[, , !is.element(info.means$FullName, to.remove$data_not_tree)]  # shape data

# Re-order to match tree tip order
tree.order <- match(aus.tree$tip.label, info.means.tree$FullName)
info.means <- info.means.tree[tree.order, ]  # metadata
mean.shapes <- mean.shapes.tree[, , tree.order]  # shape data
```

### Save intermediate data
```{r}
# Metadata and shape coordinates for all specimens
save(info, shape, file = "../Data/Processed/03-main-data.rda")

# Graphing vectors
save(genera.phylo, col.phylo, col.gen, col.unique.spp, sp.col.gen, phylo.names, pch.spp, col.unique.spp, pch.unique.spp, file = "../Data/Processed/03-graphing-vectors.rda")

# Data for phylogenetic analyses, including the tree as well as metadata and shape coordinates for mean specimens
save(info.means, mean.shapes, aus.tree, file = "../Data/Processed/03-tree-data.rda")
```
