==> Analysis/04-measure-allometry.Rmd <==
---
title: "04-measure-allometry"
author: "Ariel Marcy & Emma Sherratt"
date: "2019-01-11"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Quantifying differences in allometry
This script generates Table 1, 2, and Table S5.

Changing size is one of the most common ways that organisms can change their shape. Modifications to growth during development often have a profound impact on adult shape. The tests in this script detect how much size appears to drive shape change in our sample, relative to phylogenetic association.

Our analyses here investigate allometry at two levels:
1) Static (within-species) -- Table 1
2) Evolutionary (among clades defined by Aplin 2014 as those that radiated in Australia from distinct ancestors). -- Table 2

We do the latter, evolutionary allometric analysis with two different ANCOVAs, one without a phylogenetic reference (Table 2) and one with (Table S5). 

### Load packages, functions, and data
```{r message = FALSE}
library(geomorph)
library(RRPP)  # needed for pairwise comparisons with new `geomorph` > 3.1.0
library(stringr)
library(ape)  # phylogeny
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")

# Optional once long analyses have been run
load(file = "../Data/Processed/04-allometry-data.rda")
load(file = "../Data/Processed/04-randomization-results.rda")  
```

## 1) Measure static allometry with an ANCOVA (Procrustes ANOVA) -- reported in Table 1
```{r}
# Run static allometry test with all specimens
static.gdf <- geomorph.data.frame(coords = shape, size = info$Csize, spp = info$Taxa)
fit.static <- procD.lm(coords ~ log(size) * spp, iter = 500, data = static.gdf, print.progress = FALSE)

# Round all values in ANCOVA table to 3 sig figs & remove NAs from the table
table1 <- as.data.frame(round(fit.static$aov.table, 3))
table1[is.na(table1)] <- ""

write.table(table1, "../Data/Results/Table1_ANOVA_spp.csv", sep = ",", col.names = NA)
```

Save Regression Scores for future allometric analyses:
```{r}
static.allo.rs <- plotAllometry(fit.static, size = info$Csize, method = "RegScore")  # Allometry data
```

### 1b) Measure static allometry: Homogeneity of Slopes Test by species
Run allometry tests as well as a homogeneity of slopes test. This test may take 1-2 min to run. Updated for `geomorph` v3.1.0. 
```{r}
# HOS test defined as a function for conciseness in the next block. Returns pairwise object from RRPP for Homogeneity of Slope comparisons
TestHOS <- function(shape, size, group, iter = 500, lm.fun) {
        # Data frame
        spp.gdf <- geomorph.data.frame(coords = shape, size = size, spp = group)
        # Models
        fit.unique <- lm.fun(coords ~ log(size) * spp, iter = iter, data = spp.gdf, RRPP = FALSE)  # unique allometries by spp
        fit.common <- lm.fun(coords ~ log(size) + spp, iter = iter, data = spp.gdf, RRPP = FALSE)  # common allometries across spp (NULL HYPOTHESIS)

        return(pairwise(fit.unique, fit.common, groups = spp.gdf$spp))
        
}

# Run with HOS dataset sampled above
pw.comp <- TestHOS(shape, info$Csize, info$Taxa, lm.fun = procD.lm) 

# Count number of significant pairwise comparisons of alloemtric slope
pw.summary <- summary.pairwise(pw.comp, test.type = "VC", angle.type = "deg", confidence = 0.99)  
actual.pw <- length(which(round(pw.summary$pairwise.tables$P, 3) < 0.01))

# Export p-value matrix to Results folder
write.csv(round(pw.summary$pairwise.tables$P, 3), "../Data/Results/TableA4_HOS.csv")
```

### 1c) Randomize specimens to get NULL expantancy for pairwise significant results
We'll do the randomization procedure 500 times. This takes at least 12 hours to run, so the results are saved in the 04-randomization-results.rda file in this script.
```{r}
# Define function to shuffle species attributions and run PW comparisons
null.pairwise <- function(shape, info, iter = 500, confidence = 0.99) {
        rand.spp <- sample(info$Taxa)  # randomizes spp relative to Csize
        
        # Run the test with random sample
        pw.comp <- TestHOS(shape, info$Csize, rand.spp, lm.fun = procD.lm)
        pw.summary <- summary.pairwise(pw.comp, test.type = "VC", angle.type = "deg", confidence = confidence, show.vectors = FALSE)
        
        # Count number of significant pairwise comparisons of allometric slope
        alpha <- 1 - confidence
        return(length(which(round(pw.summary$pairwise.tables$P, 3) < alpha)))
}

# Do this randomization and testing 500 times
set.seed(42)
count.pw <- NULL
for(i in 1:500){   # 100 iterations takes ~2.5 hours
        null.pw <- null.pairwise(shape = shape, info)
        count.pw <- c(count.pw, null.pw)
        print(i)  # meter to mullify impatience
}
```

### Visualize with histogram
```{r}
# Counts are divided by 2 because pairwise table repeats each comparison
hist((count.pw[1:length(count.pw)]) / 2, main = NULL, xlab = "Significant pairwise comparisons (alpha < 0.01)")
abline(v = actual.pw / 2, lwd = 3, col = "deeppink4")  # compare random to real result above; pink for Vera
```

## 2a) Run evolutionary allometry ANCOVA -- reported in Table 2 (n = 317, clades = 7)
This analysis is directly comparable to static ANCOVA above since only the grouping variable is changed.
```{r}
# Run static allometry test with all specimens
evo.gdf <- geomorph.data.frame(coords = shape, size = info$Csize, clade = info$Clade)
fit.evo <- procD.lm(coords ~ log(size) * clade, iter = 500, data = evo.gdf, print.progress = FALSE)

# Round all values in ANCOVA table to 3 sig figs & remove NAs from the table
table2_n317 <- as.data.frame(round(fit.evo$aov.table, 3))
table2_n317[is.na(table2_n317)] <- ""

write.table(table2_n317, "../Data/Results/Table2_ANOVA_clade.csv", sep = ",", col.names = NA)
```

### 2b) Measure Evolutionary Allometry using clades and phylogeny -- Supplementary Table S5 (n = 37 species mean shapes, 8 clades)
Here we run Procrustes ANOVAs to compare the relative impact of PHYLO CLADE (Aplin 2014), size, and their interaction on shape.
```{r}
# Procrustes ANOVA with interaction
pgls.gdf <- geomorph.data.frame(coords = mean.shapes, size = info.means$MeanCsize, clade = info.means$Clade)

fit.pgls <- procD.pgls(coords ~ log(size) * clade, aus.tree, iter = 500, data = pgls.gdf, print.progress = FALSE)

# Round all values in ANCOVA table to 3 sig figs & remove NAs from the table
table2_n37 <- as.data.frame(round(fit.pgls$aov.table, 3))
table2_n37[is.na(table2_n37)] <- ""

write.table(table2_n37, "../Data/Results/TableS5_ANOVA_pgls.csv", sep = ",", col.names = NA)

# Find Reg Score values for plotting purposes later on (script 07)
evo.allo.rs <- plotAllometry(fit.pgls, size = info.means$MeanCsize, method = "RegScore")
```

### 2c) Test for power issue with species rich Pseudomys division
Rarefaction test in which 5 species from the Pseudomys division were randomly selected and the procD.pgls analysis re-run 100 times. 
```{r}
Pse.div <- info.means$FullName[which(info.means$Clade == "Pseudomys")]
count.R2 <- NULL
count.R2.p <- NULL

for (i in 1:100) {
    # Rarefy Pseudomys clade & cut down datsets accordingly
    not.in.sample <- sample(Pse.div, length(Pse.div) - 10)  # rarefy to 10 spp
    not.index <- match(not.in.sample, info.means$FullName)
    rare.info.means <- info.means[-not.index, ]  # metadata
    rare.mean.shapes <- mean.shapes[, , -not.index]  # shape data
    rare.tree <- drop.tip(aus.tree, not.in.sample)  # phylogeny
    
    # Run pgls 
    pgls.gdf.r <- geomorph.data.frame(coords = rare.mean.shapes, size = rare.info.means$MeanCsize, clade = rare.info.means$Clade)
    
    fit.pgls.r <- procD.pgls(coords ~ log(size) * clade, rare.tree, iter = 500, data = pgls.gdf.r, print.progress = FALSE)
    
    # Save R2 of interaction factor (3rd value)
    count.R2 <- c(count.R2, fit.pgls.r$aov.table$Rsq[3])
    count.R2.p <- c(count.R2.p, fit.pgls.r$aov.table$`Pr(>F)`[3])
}
```

### Visualize with histograms
Histogram of R2 result for the interaction factor (the most interesting result and only comparable statistic to other Procrustes ANOVAs). 
```{r}
setEPS()  # sets up plot export
postscript("../Data/Results/FigureR2_rare_pgls.eps")
layout(matrix(c(1,1,2,2), 2, 2, byrow = TRUE))

# Rsq value
par(mar = c(4, 4, 1.5, 0))
hist(count.R2[1:length(count.R2)], xlim = c(fit.pgls$aov.table$Rsq[3], range(count.R2)[2] + 0.01), main = NULL, xlab = "Interaction of size and clade Rsq value")
abline(v = fit.pgls$aov.table$Rsq[3], lwd = 3, col = "deeppink4")  # compare rarified results to actual result above

# P value
par(mar = c(5, 4, 1.5, 0))
hist(count.R2.p[1:length(count.R2.p)], xlim = c(fit.pgls$aov.table$`Pr(>F)`[3], range(count.R2.p)[2] + 0.01), main = NULL, xlab = "Interaction of size and clade p value")
abline(v = fit.pgls$aov.table$`Pr(>F)`[3], lwd = 3, col = "deeppink4")  # compare rarified results to actual result above

dev.off()
```

## Save intermediate data and data that takes a long time to generate
```{r}
save(fit.static, static.allo.rs, evo.allo.rs, fit.evo, fit.pgls, file = "../Data/Processed/04-allometry-data.rda")  # allometric data needed for plotting in script 08

save(count.pw, count.R2, count.R2.p, file = "../Data/Processed/04-randomization-results.rda")  # data saved from procedures that take a long time to run, namely the null hypothesis test for pairwise static allometric slope comparisons
```
