==> Analysis/05-measure-evo-rates.Rmd <==
---
title: "05-measure-evo-rates"
author: "Ariel Marcy"
date: "2/4/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Evolutionary rates
Generates Table S6: Morphological rates of evolution. 

Studies of morphological rates of evolution must take into account the phylogenetic tree because related species do not behave the same as independent samples. 

### Load packages, functions, and data
```{r message = FALSE}
library(geomorph)
library(geiger)  # for dtt test
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
```

## Compare evolutionary rates
Given the New World monkey papers (Chervaud & Marriog 2005, 2009), we would expect specialists like carnivores to have the slowest rates of evolution and rodents that evolve along the allometric line to have the greatest rates of evolution. 

First, we need to see how well the data meets the Brownian motion assumption required by the Kmult method, based on Blomberg's K (Blomberg et al. 2003), 
```{r} 
# Plot disparity through time
mean.shapes.2D <- two.d.array(mean.shapes)
aus.dtt <- dtt(aus.tree, mean.shapes.2D, nsim = 500, plot = TRUE)
```

## Morphological rates calculated and compared to each other
Groups are defined as non-specialists (1) and distinct specialists in diet and/or locomotion (2-5)
```{r}
# Define groups; 1 = those on allometric line / average body size 
groups <- rep(1, dim(info.means)[1])  # 1 is default

is.frug <- c(which(info.means$Taxa == "U_cau", info.means$Taxa == "M_gou"))
groups[is.frug] <- 2  # Define frugivores as 2

is.fol <- c(which(info.means$Taxa == "P_ora"), which(info.means$Taxa == "M_fus"), which(info.means$Taxa == "L_con"))
groups[is.fol] <- 3 # Define folivores as 3

is.carn <- c(which(info.means$Taxa == "H_chr", info.means$Taxa == "X_myo"))
groups[is.carn] <- 4 # Define carnivores as 4

groups[which(info.means$Genus == "Not")] <- 5  # Define hopping Notomys as 5

# Run analysis
names(groups) <- info.means$FullName
specialist.test <- compare.evol.rates(mean.shapes, aus.tree, groups, iter = 500, method = "simulation")
```

## Create and export Table S6
```{r}
# Evolutionary rates to 3 sig figs
rate.table <- as.matrix(round(specialist.test$sigma.d.gp.ratio, 2))
names <- c("Non-specialist", "Frugivore", "Folivore", "Carnivore", "Hopping")
colnames(rate.table) <- names
row.names(rate.table) <- names

# Put p values in lower triangle
p.adjusted <- round(p.adjust(specialist.test$pairwise.pvalue, method = "bonferroni"), 3)
rate.table[2, 1] <- p.adjusted[1]
rate.table[3, 1] <- p.adjusted[2]
rate.table[4, 1] <- p.adjusted[3]
rate.table[5, 1] <- p.adjusted[4]
rate.table[3, 2] <- p.adjusted[5]
rate.table[4, 2] <- p.adjusted[6]
rate.table[5, 2] <- p.adjusted[7]
rate.table[4, 3] <- p.adjusted[8]
rate.table[5, 3] <- p.adjusted[9]
rate.table[5, 4] <- p.adjusted[10]

# Remove zeros
for(i in 1:dim(rate.table)[1]) rate.table[i, i] <- "-"

# Export Table S6
write.table(rate.table, "../Data/Results/TableS6_EvoRates.csv", sep = ",", col.names = NA)
```
