==> Analysis/05-phylogenetic-rarefaction.Rmd <==
---
title: "05-phylogenetic-rarefaction"
author: "Ariel Marcy & Thomas Guillerme"
date: "12/6/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Phylogenetic rarefaction
These analyses addresses reviewer concerns over small sample sizes in each species bin and whether it underlies some or all of our static allometry results showing parallel slopes across almost all species. 

Phylogenetic rarefaction uses the phylogenetic tree to test whether sample size influences allometric slope measurements in our dataset. Since larger clades will have larger sample sizes, this is a powerful way to test whether rarifying all clades down to 5 specimens changed the allometric slope in a meaninful way. If all clades had a similar level of allometric slope change (measured from the observed to the rarified slopes), we could reject the hypothesis that our sample size was impacting our results.

The code below is based on Dr Thomas Guillerme's [landvR vignette here](https://github.com/TGuillerme/landvR/blob/master/inst/vignettes/rarefy_regressions.Rmd) with modifications for measuring the change in angle between two allometric trajectories. 

Generates Figure 3 and Figure S2.

### Load packages, functions, and data
```{r message = FALSE}
library(devtools)
if(!require(landvR)) install_github("TGuillerme/landvR")
library(landvR)  # get version > 0.4 for rarefaction functions
library(geomorph)
library(stringr)
library(ape)  # phylogeny
source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
load(file = "../Data/Processed/04-allometry-data.rda")

# Optional once long analyses have been run - reload to tweak plots
load(file = "../Data/Processed/05-phylo-rare-results.rda")
```

### Get the data and each clade from the phylogeny ready for the test
First, create simple data table with label, centroid size, and Reg Score (allometric shape calculated in last script)
```{r}
# Get tree tip labels into the same format as info$Taxa, i.e "G_spp"
short.label <- function(label) paste0(substr(strsplit(label, split = "_")[[1]][1], 1, 1), "_", substr(strsplit(label, split = "_")[[1]][2], 1, 3))
short.tree <- aus.tree
short.tree$tip.label <- unname(sapply(short.tree$tip.label, short.label))

# Make table
rare.table <- data.frame(info$Taxa, log(info$Csize), unname(static.allo.rs$RegScore))
colnames(rare.table) <- c("Taxa", "Csize", "RegScore")
rownames(rare.table) <- NULL
```

Next, we need to identify and isolate each element from each partition (clade), including single-species "singleton" clades using a modified version from the `ape::prop.part` function.
```{r}
clades <- prop.part.names(short.tree, singletons = TRUE)

# Make function for getting the dataset for each clade
get.data.clade <- function(clade, data, species.col.n, data.cols) {
    return(data[which(data[, species.col.n] %in% clade), data.cols])
}

# Use function to get all the datasets per clades
data.clades <- lapply(clades, get.data.clade, data = rare.table, species.col.n = 1, data.cols = c(2:3))  # 2:3 corresponds to Csize and Reg Score in rare.table (see code chunk above)
```

## Find oberved and rarified slopes for each clade
We can then calculate the delta slope (observed - rarefied slope) by using the lowest potential number of specimens in our dataset (5 - see above).
```{r}
# First define function for getting the slope of a linear model
GetSlope <- function(data) {
    return(lm(data)$coefficients[[2]])  # gives m, slope from y = mx + error
}

# Get observed slopes
obs.slopes <- unlist(lapply(data.clades, GetSlope))  # needed for the trigonometry to calculate the change in angle

# Get the rarefied statistics for DELTA SLOPES each clade with rarefy.stat, another new function in landvR
delta.slopes <- lapply(data.clades, rarefy.stat, stat.fun = GetSlope, rarefaction = 5)  # 5 = the lowest n of specimens for a species in our dataset
```

### Summarize the results
We can then summarize these results by clade size and add the clade names:
```{r}
# Clean up axis labels for plotting
new.names <- unlist(lapply(clades, paste, collapse = "+"))
for(i in 1:36) {
    if(i < 10) { 
        k <- as.character(paste("0", i, sep = "")) 
        } else k <- i
    new.names[i] <- k
}
new.names <- str_replace(new.names, "_", " ")  # removes "_" between G_spp

# Get number of spp in bins
spp.names <- unlist(lapply(clades, paste, collapse = "+"))
for(i in 1:length(spp.names))
    if(nchar(spp.names[i]) > 5) {  # detect bins of greater than 1 spp
        spp.names[i] <- ((nchar(spp.names[i]) + 1) / 6)  # gets number of spp
    }

# Combine node numbers and species number
for(i in 1:36) {
    new.names[i] <- paste(new.names[i], "; ", spp.names[i], sep = "")
}
```

## Test for "significant" delta slope angles
We converted from delta slope to delta angle using the [trigonometric equation here](https://math.stackexchange.com/questions/804089/formula-to-find-the-angle-between-two-slopes)

We set 4.5 degrees as an ideal cut-off because the absolute value of angle change can only vary from 0-90 therefore a 5% variation from 90 is 90*0.5 = 4.5 degrees.
```{r}
# Define function to test for significant delta slopes
test.delta.slope <- function(obs.slope, delta.slope, significance) {
    denom <- (1 + obs.slope * (obs.slope + delta.slope))
    angle <- abs(atan(delta.slope/denom) * 180/pi)  # converts angle into deg
    if(!missing(significance)) {
        return(ifelse(angle < significance, TRUE, FALSE))  # FALSE = failed the signifance level
    } else {
        return(angle)  # returns angle values if no significance is provided
    }
}

# Test our dataset
med.delta.slopes <- unlist(lapply(delta.slopes, median))
med.delta.slopes[which(is.na(med.delta.slopes))] <- 0  # account for NAs
is.parallel <- test.delta.slope(obs.slopes, med.delta.slopes, significance = 4.5)
sum(is.parallel)  # number of clades with less than 4.5 deg angle change

# Get delta angle change values for plotting 
angle.changes <- test.delta.slope(obs.slopes, med.delta.slopes)
```


## Plot the results
This generates Figure 3.

### Get colors to plot outliers
```{r}
# Get names & indices of outliers
out.index <- which(angle.changes > 4.5)
out.names <- new.names[angle.changes > 4.5]

# Give colors according to phylogeny
col.pts <- rep("black", length(new.names))
out.cols <- c(col.phylo["Mel"], col.phylo["Mes"], col.phylo["Not"], col.phylo["Pse"], col.phylo["Pse"], col.phylo["Rat"])

# Put outlier colors in the right places
for(i in 1:length(out.index)) {
    col.pts[out.index[i]] <- out.cols[i]
}

# Get outlier species names into "Genus species" format
tip.index <- out.index - 36  # take out indices for the 36 larger clades
long.out.names <- str_replace(aus.tree$tip.label[tip.index], "_", " ")
```

### Export Figure 3: Delta slope in CI boxplot and Delta angle as a histogram
```{r}
pdf("../Data/Results/Figure3_Phylo_Rare.pdf", family = "Times", useDingbat = FALSE)
layout(matrix(c(1,1,2,2), 2, 2, byrow = TRUE))

# 3a) Delta slope as a bloxplot
par(mar = c(1, 4, 1, 0.5))
boxplot(delta.slopes, border = "white", ylim = quantile(unlist(delta.slopes), probs = c(0.01, 0.99), na.rm = TRUE), ylab = expression(paste(Delta, " slope (95%, 50% CI)")), xlab = NULL, names = rep("", length(new.names)))  # an empty plot needed first

plot.CI(delta.slopes, type = "polygon", CI = c(95, 50), cent.tend = median, point.col = "black")
text(-0.5, 14.5, "A", cex = 1.5)
abline(h = 0, col = 'deeppink4')  # pink for Vera

# 3b) Delta angle (in degrees) as a histogram
par(mar = c(4, 4, 0, 0.5))
plot(c(1:length(med.delta.slopes)), angle.changes, xlab = "Node number; species number OR species name", ylab = expression(paste("Median ", Delta, " angle (\u00B0)", sep = "")), pch = 16, col = col.pts, ylim = c(0,90), xaxt = 'n', cex.axis = 1)
axis(side = 1, at = c(1:73), labels = new.names, las = 2, cex.axis = 0.65)

text(-0.2, 88, "B", cex = 1.5)
abline(h = 4.5, col = "deeppink4")  # significance level
text(1, 7, "4.5\u00B0 significance level", adj = c(0,0), col = "deeppink4")
legend("topright", long.out.names, col = out.cols, pch = 16, text.font = 3, cex = 0.8)

dev.off()
```

## Check the validity of the results above by randomizing species into "fake clades"
First, create some fake clade groupings (i.e. non phylogenetic) and connect it to real data by individuals
```{r}
# Getting the same number of "groups" per "clades" wihtout single spp clades
clades <- prop.part.names(aus.tree, singletons = FALSE)
n.spp.group <- sort(unlist(lapply(clades, length)), decreasing = TRUE)  # sort by big->small for graphing at the end

# Make function to get rarefied statistics for a fake clade grouping
RareFlakes <- function(n.spp.group, rare.table) {
        # Spp names for each fake clade as per n in each clade
        fake.clades <- lapply(n.spp.group, function(n.spp.group, names = rare.table$Taxa) sample(names, n.spp.group))
        
        # Use get.data.clade function from above to grab datasets for clades
        data.flakes <- lapply(fake.clades, get.data.clade, data = rare.table, species.col.n = 1, data.cols = c(2:3))
        
        # Get the rarefied statistics for each clade
        delta.slopes <- lapply(data.flakes, rarefy.stat, stat.fun = GetSlope, rarefaction = 5)
        
        return(delta.slopes)
}
```

### Apply the tests to the "fake clades" x 100
Again, calculate the delta slope (observed - rarefied slope) by using the lowest potential number of specimens in our dataset (5 - see above).
```{r}
f.delta.slopes <- list()
set.seed(802)  # Go Lake Monsters! 
for(i in 1:100) {  # Rep 100 times
        f.delta.slopes[[i]] <- do.call(cbind, RareFlakes(n.spp.group, rare.table))
}
fake.results.table <- do.call(rbind, f.delta.slopes)  # full matrix of slope values
colnames(fake.results.table) <- n.spp.group  # summarize results by clade size
``` 

### Summarize the results - Figure S2
Transform these statistics into tables to simply plot them with `boxplot`:
```{r, fig.width = 12, fig.height = 12}
fake.results.list <- apply(fake.results.table, 2, list)
colnames(fake.results.table) <- n.spp.group

pdf("../Data/Results/FigureS2_RandomClades.pdf", family = "Times", useDingbats = FALSE)

# An empty plot first
boxplot(fake.results.table, border = "white", ylim = quantile(unlist(fake.results.table), probs = c(0.01, 0.99), na.rm = TRUE), ylab = expression(paste(Delta, " slope (95%, 50% CI)")), xlab = "Number of species in random clade", names = spp.names[1:length(n.spp.group)], las = 2, cex.axis = 0.8)

# Add the results
plot.CI(unlist(fake.results.list, recursive = FALSE), type = "polygon", CI = c(50, 95), cent.tend = median, point.col = "black")
abline(h = 0, col = "deeppink4")

dev.off()
```

## Save data used for plotting
```{r}
save(new.names, spp.names, delta.slopes, med.delta.slopes, angle.changes, n.spp.group, f.delta.slopes, fake.results.table, file = "../Data/Processed/05-phylo-rare-results.rda")  # all the data needed to remake the plots
```
