==> Analysis/05-plot-evolutionary-allometry.Rmd <==
---
title: "05-plot-evolutionary-allometry"
author: "Ariel Marcy"
date: "2/14/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Phylogenetic analyses
Here we plot the Principal Components Analysis as a phylomorphospace to visualize morphological variation in context with the phylogeny. PC1 is highly correlated allometric shape. 

We also plot evolutionary allometry to visualize species which diverge.

This script generates Figure 3. 

### Load packages, functions, and data from previous steps
```{r message = FALSE}
library(stringr)
library(ape)  # needed to work with phylogenies
library(geiger)  # needed to work with phylogenies
library(phytools)  # needed to work with phylogenies
library(data.table)
library(geomorph)
library(colorspace)  # needed to darken colors programmatically
library(plotrix)  # needed to draw ellipses
load(file = "../Data/Processed/02-main-data.rda")
load(file = "../Data/Processed/03-color-data.rda")
load(file = "../Data/Processed/04-phylo-data.rda")
source("../Functions/utilities.R")  # custom functions
source("../Functions/plotGMPhyloMorphoSpace_plotmod.R")  # modified geomorph function
source("../Functions/phytools.branch.colors.R")  # modified phytools function
```

## Make multipanel Figure 3: evolutionary allometry and mean shape PCA/phylomorphosapce
First, set up colors and points by genera and important taxa
```{r}
# Make an Effective Genus column to categorize Mastacomys phylogenetically within Pseudomys
info.means.tree$EGenus <- info.means.tree$Genus
info.means.tree$EGenus[which(info.means.tree$Genus == "Mas")] <- "Pse"

# Colors for genera and different points by species within each genus; used for PCA
col.means <- PlotByGroup(info.means.tree, "EGenus", col.gen)  # standard colors from script 03 minus genus Conilurus
pch.means <- PointOutDiffSpp(info.means.tree)  # pts by unique species in a genus
pch.means[which(info.means.tree$Taxa == "M_fus")] <- 2  # open triangle

# Points to call out only specialized folivores with special characters; used for Evolutionary Allometry plot
pch.evo.allo <- rep(16, length(col.means))  # other taxa are circles
pch.evo.allo[which(info.means.tree$Taxa == "P_ora")] <- 11  # star
pch.evo.allo[which(info.means.tree$Taxa == "M_fus")] <- 2  # open triangle
```

### Run PCA
```{r}
pca.means <- plotTangentSpace(mean.shapes.tree)  # PCA of mean shapes

# Write x and y labels with proportion of variance for PC1 and PC2
PCs <- pca.means$pc.summary$importance
PC1.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
PC1.lab <- paste("PC", 1, " (", PC1.per, "%)", sep = "")
PC2.per <- round(PCs[2, 2] * 100, digits = 1)
PC2.lab <- paste("PC", 2, " (", PC2.per, "%)", sep = "")
```

### Generate data for Evolutionary Allometry plot
```{r}
# Run allometry on all 34 species
evo.procD <- geomorph.data.frame(shape = mean.shapes.tree, size = info.means.tree$MeanCsize, genus = info.means.tree$Genus)
mean.allo <- procD.allometry(shape ~ log(size), data = evo.procD)
```

### Isolate molecular rates used for phylogeny in paper and color-code branches by rates
```{r}
annotations <- aus.tree$node.comment

# Use function to strip rates from annotated NEXUS file
StripAnnotation <- function(input, annotation) {
        strsplit(strsplit(input, split = annotation)[[1]][2], split = ",")[[1]][1]
}
rates <- as.character(lapply(as.list(annotations), StripAnnotation, annotation = ",rate="))

aus.rates <- as.numeric(rates[which(!is.element(relabel.tree$tip.label, concord$tree_not_data))])
names(aus.rates) <- aus.tree$tip.label

# Plot molecular rate of evolution, light branch = slower; dark =faster
gray.branches <- phytools.branch.colors(aus.tree, aus.rates, gray)
gray.branches.10 <- darken(gray.branches, 0.1)  # make white branches visible
```

## Plot and export the multi-panel Figure 3
Evolutionary allometry plot on top and phylomorphospace on the bottom
```{r}
# Find Notomys' color
Not.col <- col.gen[which(names(col.gen) == "Not")]

setEPS()  # sets up plot export
postscript("../Data/Results/Figure3_EvoAllo_PhyloPCA.eps")
layout(matrix(c(1,1,2,2,3,3), 3, 2, byrow = TRUE))  # 3 rows, 2 columns

# Evolutionary Allometry plot
plot(x = log(info.means.tree$MeanCsize),
     y = mean.allo$Reg.proj,
     xlim = c(4.75, 6),
     col = col.means, 
     pch = pch.evo.allo,
     main = "a) Evolutionary Allometry",
     xlab = "Log centroid size", 
     ylab = "Shape (regression score)")
legend(x = "bottomright", legend = genera.phylo, col = col.phylo, pch = 16, cex = 1, pt.cex = 1, ncol = 4)

abline(lm(mean.allo$Reg.proj ~ log(info.means.tree$MeanCsize)), col = "dark grey", lwd = 1)  # plots evolutionary allometry line of best fit

text(5.77, 0.11, "Frugivores", col = "dark grey")
draw.ellipse(5.92, .115, .08, .005, angle = 1, border = "dark grey")

text(5.35, 0.08, "Folivores", col = "dark grey")
draw.ellipse(5.41, .047, .11, .02, angle = 5, border = "dark grey")

text(5.8, 0.066, "Carnivore", col = "dark grey")
text(5.05, 0.015, "Carnivore", col = "dark grey")

text(5.3, -0.055, "Hopping Notomys", col = Not.col)
draw.ellipse(5.12, -0.055, .075, .012, angle = 8, border = "dark grey")

# PCA
PCAlims = c(-0.08, 0.12)
plot(x = -pca.means$pc.scores[, 1],
     y = pca.means$pc.scores[, 2],
     xlim = PCAlims,
     ylim = PCAlims,
     col = col.means,
     pch = pch.means,
     bg = col.means,
     main = "b) Principal Component Analysis",
     xlab = PC1.lab, 
     ylab = PC2.lab)
text(.105, -0.009, "Frugivores", col = "dark grey")
draw.ellipse(.116, .015, .006, .012, angle = -12, border = "dark grey")
text(.045, 0.065, "Folivores", col = "dark grey")
draw.ellipse(.048, .035, .025, .013, angle = -19, border = "dark grey")
text(.045, -0.053, "Carnivores", col = "dark grey")
draw.ellipse(.044, -.073, .034, .007, angle = -15, border = "dark grey")
text(-.067, 0.034, "Notomys", col = Not.col)
draw.ellipse(c(-.053, -0.07), c(.014, -0.013), c(.015, .004), c(.004, .007), angle = c(-95, 0), border = c("dark grey", "dark grey"))  # draws both

# Phylomorphospace
gray.branches.10 <- rep("dark gray", length(gray.branches.10)) # STOP-GAP MEASURE to handle decision to no longer color the branches

plotGMPhyloMorphoSpace_plotmod(aus.tree, mean.shapes.tree, xaxis = -1, tip.labels = F, node.labels = F, plot.param = list(t.bg = col.means, t.pch = pch.means, t.cex = 1.5, n.cex = .75, l.col = gray.branches.10, lwd = 2, main = "c) PCA Phylomorphospace", xlab = PC1.lab, ylab = PC2.lab, xlim = PCAlims, ylim = PCAlims))
#text(0.1, 0.025, "Frugivores", col = "dark grey")
#text(0.047, 0.065, "Folivores", col = "dark grey")
#text(0.04, -0.051, "Carnivores", col = "dark grey")
#text(-0.06, 0.047, "Notomys", col = Not.col)

# Make gradient legend for molecular rates
#legend_image <- as.raster(matrix(gray.colors(10), ncol = 1))
#rasterImage(legend_image, -0.07, -0.07, -0.04, -0.044)
#text(-0.055, -0.035, "Faster rate", col = "dark grey", cex = 0.7)
#text(-0.055, -0.075, "Slower rate", col = "dark grey", cex = 0.7)

dev.off()  # finishes plot export
```
