==> Analysis/07-plot-heatmaps.Rmd <==
---
title: "07-plot-heatmaps"
author: "Thomas Guillerme & Ariel Marcy"
date: "3/11/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Plot PC1 Min/Max
Generates the heatmap images used to visualize allometric shape variation in Figure 4c,d (code to create this Figure in script 07-plot-allometry.Rmd).

To make the landmark variation heatmaps, we use Dr Thomas Guillerme's new package, `landvR` to see how specific landmarks vary along PC1.

```{r message = FALSE}
library(png)  # needed to work with screenshots
library(geomorph)
# Load libraries (and install if necessary) for landvR package
if(!require(devtools)) install.packages("devtools")
if(!require(dispRity)) install.packages("dispRity")
if(!require(landvR)) install_github("TGuillerme/landvR")
source("../Functions/utilities.R")  # loads custom functions
load(file = "../Data/Processed/03-main-data.rda")
```

### Set up
The below was adapted from a [vignette written by Dr Guillerme](https://cdn.rawgit.com/TGuillerme/landvR/8a6a6bd5/inst/vignettes/Landmark_partition_test.html).
```{r}
# Find mean configuration - i.e. mean shape in dataset
consensus <- (select.procrustes(shape, selector = mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
diff.from.mean <- coordinates.difference(coordinates = shape, reference = consensus, type = "spherical")

# Ordinate the data
twoD.shape <- two.d.array(shape)
ordination <- stats::prcomp(twoD.shape)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
gpagen.shape <- list()
gpagen.shape$coords <- shape
gpagen.shape$consensus <- consensus
class(gpagen.shape) <- "gpagen"

# Measure extremes of variation from mean on PC1
PC1.var <- variation.range(gpagen.shape, return.ID = FALSE, axis = 1, ordination = ordination, type = "spherical")

# Wrap specimens on the tangent space
wrap.PCA <- plotTangentSpace(shape, verbose = FALSE)

# Select extreme specimens (the ones that make the warp-meshes)
hypothetical.1 <- wrap.PCA$pc.shapes[[1]]
hypothetical.2 <- wrap.PCA$pc.shapes[[2]]
```

## Plot the range of variation along PC1 using a heat color scheme
Use mean shape (aka consensus) as beginning coordinate configuration.
```{r}
 # Dorsal views
open3d(windowRect = c(0, 0, 600, 600))
PC1.var.plot <- procrustes.var.plot(consensus, hypothetical.1, col = heat.colors, col.val = PC1.var[, "radius"], labels = FALSE)
view3d(phi = 0)  # flip cranium
rgl.snapshot("../Data/Processed/PC1heatmap_dor_1.png")  # save screenshot for multipanel plot

open3d(windowRect = c(0, 0, 600, 600))
PC1.var.plot <- procrustes.var.plot(consensus, hypothetical.2, col = heat.colors, col.val = PC1.var[, "radius"], labels = FALSE)
view3d(phi = 0)
rgl.snapshot("../Data/Processed/PC1heatmap_dor_2.png")

# Lateral views
open3d(windowRect = c(0, 0, 600, 600))
PC1.var.plot <- procrustes.var.plot(consensus, hypothetical.1, col = heat.colors, col.val = PC1.var[, "radius"], labels = FALSE)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/PC1heatmap_lat_1.png")

open3d(windowRect = c(0, 0, 600, 600))
PC1.var.plot <- procrustes.var.plot(consensus, hypothetical.2, col = heat.colors, col.val = PC1.var[, "radius"], labels = FALSE)
view3d(phi = 90)
rgl.snapshot("../Data/Processed/PC1heatmap_lat_2.png")
```

### Prep images for multi-panel Figure 4 (created with script 08)
This step assumes you have ImageMagik installed on your computer
```{r}
system("mogrify ../Data/Processed/PC1heatmap_*.png -trim ../Data/Processed/*.png")  # removes whitespace (requires ImageMagik)
```
