==> Analysis/08-plot-allometry.Rmd <==
---
title: "08-plot-allometry"
authors: "Ariel Marcy & Thomas Guillerme"
date: "2/14/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../allometry-rodents')
```

# Visualize allometry
This script generates Figure 1, the phylogeny colored by clade and centroid size; Figure 2, the static and common allometry plots; and Figure 4, which shows how allometry is associated with a CREA pattern using a PCA and landmark heatplots. 

CREA = Craniofacial Evolutionary Allometry (*sensu* Cardini et al. 2015): larger individuals have a longer snout and smaller relative braincase.

### Load packages, functions, and data from previous steps
```{r message = FALSE}
library(geomorph)
library(geiger)  # needed to work with phylogeny
library(phytools)
library(plotrix)  # needed to draw ellipses
library(png)
library(colorspace) 
library(stringr)
load(file = "../Data/Processed/03-main-data.rda")
load(file = "../Data/Processed/03-tree-data.rda")
load(file = "../Data/Processed/03-graphing-vectors.rda")
load(file = "../Data/Processed/04-allometry-data.rda")
source("../Functions/utilities.R")  # custom functions
source("../Functions/phytools.branch.colors.R")
```

### Prep for figure 1: Phylogeny with centroid size
Phylogeny with branch colors coordinated to centroid size and tip labels colors coordinated to major clades (see script 03).
```{r}
# Branch colors by centroid size
cs.tree <- -info.means$MeanCsize  # (-) so colors get darker with larger size
names(cs.tree) <- aus.tree$tip.label

# Tip lable colors by species, ordered in same way as tree
grp <- as.factor(str_sub(aus.tree$tip.label, 1, 3))
grp[which(grp == "Mas")] <- "Pse"
names(col.gen) <- sort(unique(grp))
col.tree <- col.gen[match(grp, names(col.gen))]
# Account for invasive Rattus
col.tree[which(aus.tree$tip.label == "Rattus_norvegicus")] <- "black"
col.tree[which(aus.tree$tip.label == "Rattus_rattus")] <- "black"

# Get edge colors
col.branch.size <- darken(phytools.branch.colors(aus.tree, -cs.tree), 0.1)  # darken so that P. delicatulus is visible

# Color the nodes of major phylogenetic groups
bg <- rep("white", 36)
bg[36] <- "black"  # invasive Rattus
bg[31] <- "dark gray"  # native Rattus
bg[29] <- "#CC1C2F"  # Hydromys group
bg[12] <- "#79ABE2"  # Pseudomys division
bg[09] <- "#004616"  # Uromys group
bg[07] <- "#A79FE1"  # Mesembriomys group

# Color the text for the phylogenetic group nodes
col.text <- rep("black", 36)
col.text[36] <- "white"  # invasive Rattus
col.text[09] <- "white"  # Uromys group

# Correct Pogonomys designation
pog.index <- which(aus.tree$tip.label == "Pogonomys_mollipilosus")
aus.tree$tip.label[pog.index] <- paste("Pogonomys_sp.")
```

## Make Figure 1
```{r}
pdf("../Data/Results/Figure1_Phylogeny_CS.pdf", family = "Times", useDingbats = FALSE)

par(mar = c(2,0,0,0))
plot.phylo(aus.tree, edge.color = col.branch.size, edge.width = 4, tip.color = col.tree, label.offset = 0.3, cex = 0.8)
axisPhylo()
nodelabels(as.character(c("01", "02", "03", "04", "05", "06", "07", "08", "09", 10:36)), frame = "circle", cex = 0.5, bg = bg, col = col.text)
mtext("                                                                                                    Millions of years ago", 1, line = 1, col = "black", adj = 0, cex = 0.8)

# Add centroid size legend
legend_image <- as.raster(matrix(c("black", gray.colors(7)), ncol = 1))
rasterImage(legend_image, 0, 2, 4, 5)
text(2.13, 5.7, "Large (~1kg)", col = "black", cex = 0.8)
text(2.13, 1.3, "Small (~10g)", col = "dark grey", cex = 0.8)

dev.off()
```

## Make simplified Figure 1 for Chapter 1 of thesis
```{r}
setEPS()
postscript("../Data/Results/Figure1_Phylogeny_Chapter1.eps")

par(mar = c(2,0,0,0))
plot.phylo(aus.tree, edge.color = "#818181", edge.width = 4, tip.color = col.tree, label.offset = 0.3, cex = 0.8)
axisPhylo()
nodelabels(as.character(c("01", "02", "03", "04", "05", "06", "07", "08", "09", 10:36)), frame = "circle", cex = 0.5, bg = bg, col = col.text)
mtext("                                                                                                  Millions of years ago", 1, line = 1, col = "black", adj = 0, cex = 0.8)

dev.off()
```

## Figure 2: Static and common allometric slopes
In _geomorph_, static allometry can be visualized with log(centroid size) plotted versus the projected regression score of shape ~ size, a common way of quantifying "size-less" shape. All static alloemtries can be compared to a common evolutionary allometry using the mean point for each static allometry:
```{r}
sa.mean.size <- aggregate(static.allo.rs$size.var, list(Taxa = info$Taxa), mean)
sa.mean.shape <- aggregate(static.allo.rs$RegScore, list(Taxa = info$Taxa), mean)
```

### Set up fancy legend to mirror phylogenetic tree
```{r}
# Extract names and get into "G. species" format
phy.long.names <- aus.tree$tip.label
genus <- paste(str_sub(phy.long.names, 1, 1), ".", sep = "")
species <- lapply(str_split(phy.long.names, "_"), '[[', 2)
spp <- str_trunc(species, 18, side = "right", ellipsis = ".")  # shorten spp names
legend.names <- str_c(genus, spp[], sep = " ")  # combine

# Order like phylogeny by getting names to match phylo.names format
first <- str_sub(phy.long.names, 1, 1)
second <- str_sub(lapply(species, '[[', 1), 1, 3)
match.names <- str_c(first, second, sep = " ")

# Subset legend vectors to exclude Z ped which is not in the phylogeny
leg.names <- legend.names[match(phylo.names[-which(phylo.names == "Z ped")], match.names)]  # also ordered with phylogeny
col.legend <- col.unique.spp[-which(phylo.names == "Z ped")]
pch.legend <- pch.unique.spp[-which(phylo.names == "Z ped")]

# Fix Pogonomys sp. 
leg.names[is.na(leg.names)] <- "Pogonomys sp." 
```

### Plot/export static and evolutionary allometry multi-panel Figure 2
Here we consolidate the information above into two static allometry plots, one with all the data points and one with just the regression "pred-lines" for each species to compare with the evolutionary allometry of the entire group. 
```{r}
pdf("../Data/Results/Figure2_Static_Allo.pdf", family = "Times", useDingbats = FALSE)

layout(matrix(c(1,1,2,2), 2, 2, byrow = TRUE))

# Plot of all data points with evolutionary allometry
par(mar = c(1, 4, 1, 0))
plot(fit.static,
     type = "regression",
     reg.type = "RegScore",
     predictor = log(info$Csize),
     xlim = c(4.7, 6.4),
     col = sp.col.gen, 
     pch = pch.spp,
     bg = sp.col.gen,
     xlab = NA,
     xann = 'n',
     ylab = "Shape (Regression Score)",
     bty = "n")
text(4.7, 0.048, "A", cex = 1.5)
abline(lm(sa.mean.shape$V1 ~ sa.mean.size$x), col = "dark grey", lwd = 2)  # plots evolutionary allometry
r.stat <- round(fit.static$aov.table$Rsq[1], 2)
text(5.5, -0.03, bquote(R^2 ~ "of log(size) = " ~ .(r.stat)), cex = 1.1, col = "black")

# Species Legend
legend(6.05, 0.05, legend = leg.names[1:20], col = col.legend[1:20], border = NULL, pch = pch.legend[1:20], pt.bg = col.legend[1:20], cex = 0.85, ncol = 1, bg = "white", bty = "n", text.font = 3)

# Plot of static allometry with evolutionary allometry
par(mar = c(5, 4, 1.5, 0))
plot(fit.static,
     type = "regression",
     reg.type = "PredLine",
     predictor = log(info$Csize),
     xlim = c(4.7, 6.4),
     col = sp.col.gen, 
     pch = pch.spp,
     bg = sp.col.gen,
     xlab = "Log(centroid size)",
     ylab = "Shape (Predicted Value)",
     bty = "n")
text(4.7, 0.11, "B", cex = 1.5)
r.stat3 <- round(fit.static$aov.table$Rsq[3], 2)
text(5.6, -0.05, bquote(R^2 ~ "of log(size):species = " ~ .(r.stat3)), cex = 1.1, col = "black")

# Species Legend
legend(6.05, 0.147, legend = leg.names[21:37], col = col.legend[21:37], border = NULL, pch = pch.legend[21:37], pt.bg = col.legend[21:37], cex = 0.85, ncol = 1, bg = "white", bty = "n", text.font = 3)

dev.off()
```

## Figure 4: multi-panel evolutionary allometry and mean shape PCA
First, set up colors and points by genera and important taxa
```{r}
# Colors for genera and distinct points by species within each genus for PCA
col.means <- PlotByGroup(info.means, "EGenus", col.gen)  # standard colors from script 03 minus genus Conilurus
pch.means <- PointOutDiffSpp(info.means$EGenus)  # pts by unique spp in genus
pch.means[which(info.means$Taxa == "M_fus")] <- 2  # open triangle

# Points to call out only specialized folivores with special characters; used for Evolutionary Allometry plot
pch.evo.allo <- rep(16, length(col.means))  # other taxa are circles
pch.evo.allo[which(info.means$Taxa == "P_ora")] <- 10  # quartered circle
pch.evo.allo[which(info.means$Taxa == "M_fus")] <- 2  # open triangle

# Legend that include Mastacomys in Fig 4A
gen.phylo.new <- c(genera.phylo[1:7], "Mas", genera.phylo[8:length(genera.phylo)])  # legend words

col.phylo.new <- c(col.phylo[1:7], col.phylo[7], col.phylo[8:length(col.phylo)])  # legend colors

pch.phylo.new <- rep(16, length(col.phylo.new))
pch.phylo.new[8] <- 2  # open triangle
```

### Run PCA
```{r}
pca.means <- plotTangentSpace(mean.shapes)  # PCA of mean shapes

# Write x and y labels with proportion of variance for PC1 and PC2
PCs <- pca.means$pc.summary$importance
PC1.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
PC1.lab <- paste("PC", 1, " (", PC1.per, "%)", sep = "")
PC2.per <- round(PCs[2, 2] * 100, digits = 1)
PC2.lab <- paste("PC", 2, " (", PC2.per, "%)", sep = "")
```

### Read in heatmap pngs created with script 07
```{r}
dor1 <- readPNG("../Data/Processed/PC1heatmap_dor_1.png")
dor2 <- readPNG("../Data/Processed/PC1heatmap_dor_2.png")
lat1 <- readPNG("../Data/Processed/PC1heatmap_lat_1.png")
lat2 <- readPNG("../Data/Processed/PC1heatmap_lat_2.png")
```

## Plot and export the multi-panel Figure 4
Evolutionary allometry plot on top, PCA in the middle, heatmaps on the bottom.
```{r}
# Set up layout and export settings
pdf("../Data/Results/Figure4_Allo_Shape.pdf", family = "Times", useDingbats = FALSE)
mat <- matrix(c(1,1,2,2:6), 4, 2, byrow = TRUE)  # 6 plots, 4 rows, 2 cols
layout(mat, widths = rep.int(1, ncol(mat)), heights = rep.int(1, nrow(mat))) 

# 1) Evolutionary Allometry plot
par(mar = c(4, 4, 1, 1))  # sets the margins
plot(x = log(info.means$MeanCsize),
     y = evo.allo.rs$RegScore,
     xlim = c(4.75, 6),
     col = col.means,
     pch = pch.evo.allo,
     xlab = "Log centroid size", 
     ylab = "Shape",
     bty = "n")
text(4.75, 0.08, "A", cex = 1.5)
legend("bottomright", inset = c(0.026, 0), legend = gen.phylo.new, col = col.phylo.new, pch = pch.phylo.new, cex = 0.9, pt.cex = 0.95, ncol = 3, bty = "n", text.font = 3)

abline(lm(evo.allo.rs$RegScore ~ log(info.means$MeanCsize)), col = "dark grey", lwd = 1)  # plots evolutionary allometry line of best fit

# Frugivores
text(5.985, evo.allo.rs$RegScore[which(info.means$Taxa == "M_gou")] + 0.012, pos = 2, "Frugivores", col = "grey17")
#draw.ellipse(5.913, .065, .06, .01, angle = -2, border = "dark grey")

# Folivores
text(log(info.means$MeanCsize)[which(info.means$Taxa == "M_fus")], evo.allo.rs$RegScore[which(info.means$Taxa == "M_fus")] - 0.005, "Folivores", pos = 2, col = col.gen[which(names(col.gen) == "Pse")])  # color of Pse

text(log(info.means$MeanCsize)[which(info.means$Taxa == "L_con")] + 0.1, evo.allo.rs$RegScore[which(info.means$Taxa == "L_con")] - 0.01, "Folivore", pos = 2, col = col.gen[which(names(col.gen) == "Lep")])  # color of Pse

# Carnivores
text(log(info.means$MeanCsize)[which(info.means$Taxa == "X_myo")], evo.allo.rs$RegScore[which(info.means$Taxa == "X_myo")], pos = 2, "Carnivore", col = col.gen[which(names(col.gen) == "Xer")])  # Xeromys
text(log(info.means$MeanCsize)[which(info.means$Taxa == "H_chr")], evo.allo.rs$RegScore[which(info.means$Taxa == "H_chr")] - 0.005, pos = 2, "Carnivore", col = col.gen[which(names(col.gen) == "Hyd")])  # Hydromys

# Hopping Notomys
text(log(info.means$MeanCsize)[which(info.means$Taxa == "N_mit")] + 0.03, evo.allo.rs$RegScore[which(info.means$Taxa == "N_mit")], pos = 4, expression(paste("Hopping ", italic("Notomys"))), col = col.gen[which(names(col.gen) == "Not")])
draw.ellipse(5.12, -0.035, .075, .009, angle = 2, border = "dark grey")

# 2) PCA
par(mar=c(4, 4, 1, 1))  # sets the margins
PCAlims = c(-0.08, 0.12)
plot(x = -pca.means$pc.scores[, 1],
     y = pca.means$pc.scores[, 2],
     xlim = PCAlims,
     ylim = PCAlims,
     col = col.means,
     pch = pch.means,
     bg = col.means,
     xlab = PC1.lab, 
     ylab = PC2.lab,
     bty = "n")
text(-0.08, 0.1, "B", cex = 1.5)
text(.112, -0.01, "Frugivores", col = "grey17")
text(.05, 0.065, "Folivores", col = col.gen[which(names(col.gen) == "Pse")])
text(.034, 0, "Folivore", col = col.gen[which(names(col.gen) == "Lep")])
draw.ellipse(.048, .032, .022, .013, angle = -19, border = "dark grey")
text(.03, -0.063, "Carnivore", col = col.gen[which(names(col.gen) == "Xer")])
text(.06, -0.063, "Carnivore", col = col.gen[which(names(col.gen) == "Hyd")])
text(-.066, 0.025, expression(italic(Notomys)), col = col.gen[which(names(col.gen) == "Not")])
draw.ellipse(c(-.053, -0.07), c(.014, -0.012), c(.015, .0035), c(.004, .008), angle = c(-95, 0), border = c("dark grey", "dark grey"))  # draws both

# 3) Plot dorsal view - mean to min
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(dor1)[2]), c(0, dim(dor1)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(dor1, 0, 0, dim(dor1)[2], dim(dor1)[1])
text(0, dim(dor1)[1]-10, "C", cex = 1.5)

# 4) Plot dorsal view with same dimensions as above - mean to max
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(dor1)[2]), c(0, dim(dor1)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE) 
rasterImage(dor2, 0, 0, dim(dor1)[2], dim(dor1)[1])
text(0, dim(dor1)[1]-10, "E", cex = 1.5)

# 5) Plot lateral view  - mean to min
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(lat1)[2]), c(0, dim(lat1)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE)
rasterImage(lat1, 0, 0, dim(lat1)[2], dim(lat1)[1])
text(0, dim(lat1)[1]-10, "D", cex = 1.5)

# 5) Plot lateral view with same dimensions as above - mean to max
par(mar = c(0, 0, 1, 0))
plot(c(0, dim(lat1)[2]), c(0, dim(lat1)[1]), type = "n", axes = FALSE, xlab = "", ylab = "", asp = TRUE)
rasterImage(lat2, 0, 0, dim(lat1)[2], dim(lat1)[1])
text(0, dim(lat1)[1]-10, "F", cex = 1.5)

# Make gradient legend
legend_image <- as.raster(matrix(heat.colors(20), ncol = 1))
rasterImage(legend_image, 440, 150, 453, 210)
text(387, 198, "High variation", col = "grey55", cex = 1)
text(419, 160, "Low", col = "grey55", cex = 1)

dev.off()
```
