==> Analysis/12-calculate-user-error.Rmd <==
---
title: "12-calculate-user-error"
author: "Ariel Marcy"
date: "9/26/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../aus-rodent-skulls')
```
# Calculate user error
Geometric morphometric studies usually require a lot of user-generated data and it's important to assess how much user error is created during the landmarking process. 

First, we visualize whether any outliers can be explained by mistakes in landmark placement (e.g. a landmark "falling through" the nasal to land on the inner basicranium).

Second, we compare the shape variation between 20 individuals, from a variety of different species, landmarked twice to calculate the intra-observer (i.e. digitizer) error.

### Load packages, functions, and data from previous steps
```{r message = FALSE}
library(stringr)
library(stringi)
library(data.table)
library(geomorph)
source("../Functions/utilities.R")  # custom functions defined in utilities.R
load(file = "../Data/Processed/01-bilat-big-data-w-ERR.rda")
load(file = "../Data/Processed/01-bilat-sm-data-w-ERR.rda")
```

## 1) Check for outliers and fix landmarking errors
_geomorph_ comes with a handy function to check for outliers. Since our sample has a wide range of species, we decided to include the groups factor. This means `plotOutliers()` returns likely outliers for each genus instead of the group as a whole. When groups were not defined, only large species were returned as "outliers". 
```{r}
# Plot outliers by genus for big patch dataset
outliers <- plotOutliers(big.bY.shape, groups = big.sp.info$Genus, inspect.outliers = T)  # try changing to Taxa??
```

Run again for the small patch protocol dataset:
```{r}
outliers <- plotOutliers(sm.bY.shape, groups = sm.sp.info$Genus, inspect.outliers = T)
```

### Visualize why they are outliers by looking at landmark vectors
We can examine the landmark placements in outliers as compared to the specimen closest to the median, based on the plots above. This allows us to determine if there were any major landmarking errors that might impact further analyses.

Genus abbreviatons and number of outliers were inserted where needed for each genus. This code was only used once: after all landmarks were captured and before the other major analyses were completed.

We wrote a function, `MatchSpecShape()` to help us visualize outlier shape information. The function returns the 3D shape of specimen of interest. We can use it to compare the outliers to the median specimen in its genus.
```{r}
# Get shape data for outliers indicated by plots above
out.1 <- MatchSpecShape(outliers$Mel[1], sm.sp.info, sm.bY.shape)  # 1st outlier in genus Mel
out.2 <- MatchSpecShape(outliers$Mel[2], sm.sp.info, sm.bY.shape)  # 2nd outlier in genus Mel

# Compare outlier specimens to median specimen
med.spec <- outliers$Mel[median(outliers$Mel)]  # median specimen in Mel
med.shape <- MatchSpecShape(med.spec, sm.sp.info, sm.bY.shape)

# Compare landmark locations in outlier specimens to median specimen
plotRefToTarget(med.shape, out.1, method = "vector", label = FALSE)
plotRefToTarget(med.shape, out.2, method = "vector", label = FALSE)
```
**Repeat the above for each genus/patch protocol as needed. As landmarking errors are found, fix in Viewbox, export new coordinates, and begin again from script 01.** 

## 2) Calculate intra-observer error
Now that the shape data have been processed via GPA + bilateral symmetry and corrected for obvious landmarking errors, we can separate the data intended to measure user error from the main data for the remaining analyses. 

Make the datasets to be used for the rest of the analyses.
```{r}
main <- which(big.sp.info$Rep == "non")
big.shape <- big.bY.shape[, , main]
big.info <- big.sp.info[main, ]

sm.shape <- sm.bY.shape[, , main]  # main is the same for big and small patch protocols, tested with all.equal()
sm.info <- sm.sp.info[main, ]
```

Create an error measurement dataset for both protocols that includes replicates and the specimens they duplicated (i.e. each CatNum should have two entries, one labeled "ERR", and one labled "non"). These data will be used to calculate user error. 
```{r}
# Find index for both versions of specimens with replicates using rep.spec list make in script 01.
rep.CatNums <- as.character(str_sub(rep.spec, 9, -1))  # isolate CatNum
error.data <- which(sm.sp.info$CatNum %in% rep.CatNums)  # error.data is the same for small and big patch protocols, tested with all.equal()

# Subset shape and metadata tables for each protcol
big.err.shape <- big.bY.shape[, , error.data]
big.err.info <- big.sp.info[error.data, ]

sm.err.shape <- sm.bY.shape[, , error.data]
sm.err.info <- sm.sp.info[error.data, ]
```

### Calculate user error aka Repeatability
Here, we calculate repeatability for each patch protocol from formulas 1-3 in Fruciano 2016. The function `RepAbility()` was modified from Marcy et al. 2018.
```{r}
# Note: n.rep = 2 in this case because we measured each specimen twice
RepAbility(big.err.shape, big.err.info$CatNum, n.Rep = 2)  
RepAbility(sm.err.shape, sm.err.info$CatNum, n.Rep = 2)
```

### Calculate repeatability without patch points
Patch points could artificially inflate our repeatability numbers since they are placed automatically by the Viewbox software. To test this, we calculated repeatability for each protocol with patch points removed. 

Note that removing patch points will remove 16 more points from the big patch protocol than from the small patch protocol. 
```{r}
# Find patch numbers for each protocol
big.patches <- str_detect(big.pt.names, "PAT")
big.pat.num <- which(big.patches == TRUE)

sm.patches <- str_detect(sm.pt.names, "PAT")
sm.pat.num <- which(sm.patches == TRUE)

# Subset the shape datasets for only fixed LMs and curve semi-LMs
big.patchless.shape <- big.err.shape[-big.pat.num, , ] 
sm.patchless.shape <- sm.err.shape[-sm.pat.num, , ]

# Re-run repeatability
RepAbility(big.patchless.shape, big.err.info$CatNum, n.Rep = 2)
RepAbility(sm.patchless.shape, sm.err.info$CatNum, n.Rep = 2)
```
Taking out patch points decreases the repeatability by less than 2% for each protocol:
Big Patch is 93.0% with patches, 91.3% without patches (140 points removed)
Small Patch: 93.0% with patches, 91.5% without patches (124 points removed) 

### Save intermediate data 
These are the main datasets for both protocols without replicates
```{r}
save(big.shape, big.info, file = "../Data/Processed/02-main-data-big-protocol.rda")
save(sm.shape, sm.info, file = "../Data/Processed/02-main-data-sm-protocol.rda")
```
