==> Analysis/16-run-phylogenetic-analyses.Rmd <==
---
title: "16-run-phylogenetic-analyses"
author: "Ariel Marcy"
date: "11/19/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../aus-rodent-skulls')
```

# Phylogenetic analyses
Evolutionary studies must take into account the phylogenetic tree because related species do not behave the same as independent samples. 

Many of these steps are based on Dr Meriam Zelditch et al.'s A Practical Companion (2012). 
> Zelditch ML Swiderski D Sheets HD. 2012. A practical companion to geometric morphometrics for biologists: running analyses in freely-available software. Available at: http://booksite.elsevier.com/9780123869036.

### Load packages, functions, and data
```{r message = FALSE}
library(stringr)
library(data.table)
library(geomorph)
library(ape)  # needed to work with phylogenies from here on
library(geiger)
library(nlme)
library(caper)
library(car)
source("../Functions/utilities.R")  # loads custom functions
load(file = "../Data/Processed/03-sm-PCA-with-color.rda")
load(file = "../Data/Processed/04-sm-data-with-traits.rda")
```

### Rename shape array to match phylogeny
First, we'll rename the dimnames for the shape dataset so that it's only "Genus_species". This is so we can re-order the shape data to match the tree later on. (We can still reference the CatNum in the metadata.)
```{r}
shape.phylo <- shape  # initiate shape optimized for phylogenetics
dimnames(shape.phylo)[[3]] <- as.list(paste(info.traits$Genus.y, info.traits$Species.y, sep = "_"))
```

## Calculate mean shape and centroid size for each species
This step has to be done before working with the tree. To calculate the mean shape for each species, we'll use a function from `geomorph` called `m.shape()`. 

To get the means, we'll subset for each unique species. The mean shape function, `mshape()` gives a 2D array as an output. Therefore, the `for` loop below stores mean shapes in a 2D array format recognized by `geomorph`: (n x [p x k]), where n is number of specimens, p is number of landmark points, and k is number of dimensions. Then we convert to a 3D array with format (p x k x n) since a 3D array is required by most `geomorph` functions. 
```{r}
info.traits$FullName <- paste(info.traits$Genus.y, info.traits$Species.y, sep = "_")  # give names to metadata in the same format as the tree "Genus_species"
full.names <- unique(info.traits$FullName)

# For loop to subset shape data by species and calculate mean shape
mean.shapes.2D <- NULL  # initiate blank data frame
for (i in 1:length(full.names)) {  
        is.taxa <- which(info.traits$FullName == full.names[i])
        shape.for.mean.2D <- shape.phylo[, , is.taxa]
        mean.shapes.2D <- rbind(mean.shapes.2D, mshape(shape.for.mean.2D))  # mshape() finds mean specimen; rbind() is critical, output cols must = 3 for 3D
}
mean.shapes <- arrayspecs(mean.shapes.2D, 325, 3)  # convert to 3D array

# Make names match the tree, make metadata match mean shape data (alphabetical)
dimnames(mean.shapes)[[3]] <- full.names  # shape names match tree names
info.means <- info.traits[!duplicated(info.traits$Taxa), -c(4, 5)]  # make metadata; remove columns that don't make sense: CatNum and Order
info.means <- info.means[order(info.means$FullName), ]  # order same as shape's
```

### Calculate mean centroid size
We need to update the centoid size column for info.means to contain the *mean* centroid size by taxa. Centroid size is stored in the info.traits metadata table.
```{r}
# Calculate mean centroid size and store in info.means
cs.means <- aggregate(info.traits[, 6], list(Taxa = info.traits$FullName), mean)
info.means$Csize <- cs.means[, 2]  # store in vector; aggregate alphabetized it
names(info.means)[names(info.means) == 'Csize'] <- 'MeanCsize'  # rename column

# Sort shape by alphabetical order, too
mean.shapes <- mean.shapes[ , , sort(dimnames(mean.shapes)[[3]])]
```

### Plot PCA of skull shape means by species
Before adding the tree, we can plot the PCA of mean shapes, which will contain more species than the set subsetted by species found in the tree.
```{r}
# Set colors for genera and different points for species within each genus
col.means <- PlotByGroup(info.means, "EGenus", col.gen)  # standard colors
pch.means <- PointOutDiffSpp(info.means)  # points by unique species in a genus

# Plot PCA of mean shapes
pca.means <- plotTangentSpace(mean.shapes)
PlotPCA(pca.means, 1, 2, col.means, pch.means, flip.axis1 = TRUE)
legend(0.142, 0.047, legend = info.means$Taxa, col = col.means, border = NULL, pch = pch.means, pt.bg = col.means, cex = 0.305)
```

### Correlation of PC1 and PC2 wtih Centroid size
Here we find Pearson's r for correlation with PC1 and PC2.
```{r}
cor.1 <- cor.test(info.means$MeanCsize, -pca.means$pc.scores[, 1], method = "pearson")  # remember to flip axis
cor.1  # PC1
cor.2 <- cor.test(info.means$MeanCsize, pca.means$pc.scores[, 2], method = "pearson")  
cor.2  # PC2
```

## Prepare the tree data for phylogenetic analyses
The commands in `geomorph` require the tree and datasets to have exactly the same set of species.

### Load in phylogenetic tree
The most recent tree for Australian rodents is from Smissen & Rowe 2018, Figure 2. The tree is a fossil calibrated ultrametric tree from BEAST2 for Hydromini, which includes all the Australo-Papuan old endemics in Murinae.
```{r}
Rowe.tree <- "../Data/Processed/Smissen-Rowe-2018-concat.tre"  # 
raw.tree <- read.tree(Rowe.tree)
```

### Extract only Genus and Species names from the tree
The tree file has uneven amounts of information for each entry, some entries have catalog numbers and others do not. Therefore we needed a `for` loop to grab the first two elements from a list, which always corresponded to Genus and Species.

We reassign only Genus and Species names (in form "Genus_species") to the tree so that going forward it is easier to compare to the shape data. 
```{r}
no.cat.tree <- raw.tree  # initiate new, modified tree with no CatNums
label.list <- strsplit(raw.tree$tip.label, "_")  # separate info into list

# For loop finds Genus & Species and reassigns tip name as "Genus_species"
for (i in 1:length(label.list)) {
        label <- unlist(label.list[i])
        no.cat.tree$tip.label[i] <- paste(label[1], label[2], sep = "_")
}
```

### Rename nodes
Since the 2018 Smissen & Rowe tree focused primarily on New Guinae species, it does not have all of the Australian species in our shape dataset. However, it does include sister species with the same relative branch length to other genera in the study. Therefore, we decided to rename some of the nodes with Australian species. 

**Renamed nodes:**
_Pogonomys macruourus_ (NG) to _Pogonomys mollipilosus_ (Aus)
_Melomys bannisteri_ (NG) to _Melomys burtoni_ (Aus)

```{r}
renamed.tree <- no.cat.tree
P.mac <- which(str_detect(no.cat.tree$tip.label, "Pogonomys_macrourus"))
renamed.tree$tip.label[P.mac] <- paste("Pogonomys_mollipilosis")

M.ban <- which(str_detect(no.cat.tree$tip.label, "Melomys_bannisteri"))
renamed.tree$tip.label[M.ban] <- paste("Melomys_burtoni")

# Optional plot
plot(renamed.tree)
tiplabels(cex = 0.1)
nodelabels(cex = 0.1)
axisPhylo()
```

### Prune the tree and shape datasets to match each other
Now that the names in the Smissen & Rowe tree can be compared to our info.traits metadata, we can shape dataset to have all the same species.

First, prune the tree of species not in the shape data
```{r}
concord <- name.check(renamed.tree, renamed.tree$tip.label, dimnames(mean.shapes)[[3]])  # check concordance
aus.tree <- drop.tip(renamed.tree, concord$tree_not_data)  # prune tree to Aus
```

Second, remove duplicates of species in the tree
```{r}
dupes <- which(duplicated(aus.tree$tip.label))
aus.tree <- drop.tip(aus.tree, dupes)
```

Third, subset and re-order the metadata and shape datasets to match the order of species found in the tree. 
```{r}
# Subset
to.remove <- name.check(aus.tree, aus.tree$tip.label, dimnames(mean.shapes)[[3]])
info.means.tree <- info.means[!is.element(info.means$FullName, to.remove$data_not_tree), ]  # metadata
mean.shapes.tree <- mean.shapes[, , !is.element(info.means$FullName, to.remove$data_not_tree)]  # shape data

# Re-order to match tree tip order
tree.order <- match(aus.tree$tip.label, info.means.tree$FullName)
info.means.tree <- info.means.tree[tree.order, ]  # metadata
mean.shapes.tree <- mean.shapes.tree[, , tree.order]  # shape data
```

## Phylogenetic Analyses
First, test for phylogenetic signal in the shape data. 
```{r}
physignal(mean.shapes.tree, aus.tree)
```

### Evolutionary allometry:
```{r}
# Centroid size only
evo.allo.gdf <- geomorph.data.frame(shape = mean.shapes.tree, CS = info.means.tree$MeanCsize)
evo.allo <- procD.pgls(shape ~CS, aus.tree, data = evo.allo.gdf)
evo.allo

# Interaction with genus
evo.allo.gen.gdf <- geomorph.data.frame(shape = mean.shapes.tree, CS = info.means.tree$MeanCsize, genus = info.means.tree$EGenus)
evo.allo.gen <- procD.pgls(shape ~CS * genus, aus.tree, data = evo.allo.gen.gdf)
evo.allo.gen

# Example of a test of homogeneity of slopes, plus pairwise slopes comparisons
HOS <- procD.pgls(f1 = shape ~ log(CS) * genus, aus.tree, data = evo.allo.gen.gdf)
summary(HOS, formula = FALSE) # formulas too long to print


```

**Historical Inference: Evolutionary transitions in size and shape**
First try with mean shapes:
```{r}
ancestral.shapes <- plotGMPhyloMorphoSpace(aus.tree,
mean.shapes.tree, tip.labels = T, node.labels = F)
```

Next, try with "size-less" residuals of mean shape:
```{r}
ancestral.res <- plotGMPhyloMorphoSpace(aus.tree,
evo.allo$pgls.residuals, tip.labels = T, node.labels = F)
```

### Compare evolutionary rates
Given the new world monkey papers (Chervaud & Marriog 2005, 2009?), we would expect the carnivores to have the slowest rates of evolution and rodents, which evolve along the allometric line to have the greatest rates of evolution. 

First, we need to make sure the data meets the brownian motion assumption required by the Kmult method, based on Blomberg's K (Blomberg et al. 2003), 
```{r} 
### NOT WORKING YET ###
mean.shapes.tree.2D <- two.d.array(mean.shapes.tree)
aus.dtt <- dtt(mean.shapes.tree.2D, aus.tree, plot = TRUE)
```

**_Notomys_ vs Pseudomys vs AOEs vs Other**
```{r}
# Set up grouping factor
gp.notomys <- rep(0, dim(info.means.tree)[1])  # set "other" as 0
gp.notomys[which(info.means.tree$Wave == 1)] <- 1  # set AOEs to 1
gp.notomys[which(info.means.tree$Genus == "Not")] <- 2  # set Notomys
gp.notomys[which(info.means.tree$EGenus == "Pse")] <- 3  # set Pseudomys
names(gp.notomys) <- info.means.tree$FullName  # required to run analysis

# Run analysis
compare.evol.rates(mean.shapes.tree, aus.tree, gp.notomys)
```

**Carnivores vs non-carnivores**
```{r}
is.carnivore <- as.numeric(info.means.tree$Feeding == "Carnivorous")
names(is.carnivore) <- info.means.tree$FullName
compare.evol.rates(mean.shapes.tree, aus.tree, is.carnivore)

# Carnivores vs waves
carn.vs.waves <- info.means.tree$Wave - is.carnivore  # makes carnivores = 0
names(carn.vs.waves) <- info.means.tree$FullName
compare.evol.rates(mean.shapes.tree, aus.tree, carn.vs.waves)
```
Neither test showed a significant difference in evolutionary rates. 

**Carnivores vs herbivores**
Mastacomys fuscus and Pseudomys oralis are the most specialized herbivores.
```{r}
gp.cvh <- as.numeric(info.means.tree$Feeding == "Carnivorous")
gp.cvh[which(info.means.tree$Taxa == "M_fus")] <- 2
gp.cvh[which(info.means.tree$Taxa == "P_ora")] <- 2
names(gp.cvh) <- info.means.tree$FullName
compare.evol.rates(mean.shapes.tree, aus.tree, gp.cvh)
```

### Compare evolutionary rates of centroid size by wave
```{r}
Csize <- matrix(info.means.tree$MeanCsize, dimnames = list(info.means.tree$FullName))  # make matrix Csize with names
gp.wave <- info.means.tree$Wave
gp.wave[length(gp.wave)] <- 3  # group Mus with invasive Rattus
names(gp.wave) <- info.means.tree$FullName
compare.evol.rates(Csize, aus.tree, gp = gp.wave, iter = 49)

```
AOEs have the highest rate of centroid size evolution, especially compared to _Rattus_.
